### 2025-10-07 #9154 @editorialbot
[PRE REVIEW]: SpinVibe: Simulating spin-phonon coupling for molecular qubits
**Submitting author:** @victorchanglee (<a href="http://orcid.org/0000-0003-1767-6229">Victor Chang Lee</a>)
**Repository:** https://github.com/victorchanglee/SpinVibe
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
