### 2025-10-07 #9154 @kyleniemeyer
Your submission looks like it could be in scope in terms of research software, but unfortunately it is has some larger issues based on our [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) that require some work before we proceed:
- It is missing an open source license
- It is not structured properly to be an installable Python package (see Python packaging guidelines at https://packaging.python.org/)
- there is no documentation
- there are no tests

Please let me know if you are able and willing to address these issues in the near term, and I can pause your submission. If you are not able, your submission will be rejected for not meeting some critical JOSS submission requirements.
