### 2025-10-07 bf43c24
Delete spin_phonon.py
