+++ spin_phonon.py
@@ -1,407 +1,0 @@
-import numpy as np
-import hamiltonian
-import coupling
-import math_func
-import redfield
-import measure
-import h5py as h5
-from constants import Bohrmagneton, k_B
-import time
-from mpi4py import MPI
-import scipy.linalg
-from datetime import datetime
-
-class spin_phonon:
-    def __init__(self, B, S, Ncells, Delta_alpha_q, rot_mat, pol, T, tf, dt, file_reader,save_file,init_type='polarized',R_type=None):
-        
-        comm = MPI.COMM_WORLD
-        rank = comm.Get_rank()
-        size = comm.Get_size()
-        
-        init_time = time.perf_counter()
-
-        if rank == 0:
-            print(datetime.now().strftime("%Y-%m-%d %H:%M:%S"))
-            print("---------------------------------------")
-            print("|Start Spin-phonon coupling simulation|")
-            print("---------------------------------------")
-            print("\n")
-
-        timer_input = time.perf_counter()
-        
-        # Parameter initialization
-        self.B_T = B  # Magnetic field vector
-        self.Delta_alpha_q = Delta_alpha_q  # Broadening parameter
-        self.S = S  # Spin quantum number 
-        self.rot_mat = rot_mat
-        self.pol = pol  # Polarization vector
-        self.T = T  # Temperature in Kelvin
-        self.init_type = init_type
-        self.R_type = R_type
-        self.Ncells = Ncells  # Number of unit cells 
-        self.save_file = save_file
-        if rank == 0:
-            print("Input Parameters")
-            print("================")
-            print("Magnetic field:", self.B_T)
-            print("S:", self.S)
-            print("T:", self.T)
-            print("Number of unit cells:", self.Ncells)
-            print("Broadening:", self.Delta_alpha_q)
-            print("Population type:", self.init_type)
-            print("Polarization:", self.pol)
-            print("Rotational matrix:")
-            print(self.rot_mat)
-            print("\n")
-
-            
-
-        self.B = self.B_T * Bohrmagneton  # Convert from T to cm-1
-        self.m = np.arange(-self.S, self.S+1, 1)
-        self.hdim = int(2*self.S + 1) 
-
-        # File reading (ALL ranks need this data)
-        self.file_reader = file_reader
-        self.q_vector, self.omega_q, self.L_vectors = self.file_reader.read_phonons()
-        self.R_vectors, self.reciprocal_vectors = self.file_reader.read_atoms()
-        self.q_vector = self.q_vector @ self.reciprocal_vectors # Convert q vectors to A^-1
-
-        # More parameter setup (ALL ranks)
-        self.N_atoms = self.R_vectors.shape[0]  # Number of atoms
-        self.Nomega = len(self.q_vector)  # Number of phonon modes
-        self.Nq = self.q_vector.shape[0]  # Number of q points
-        
-        g_tensor, d_tensor = self.file_reader.read_orca() #Read g-matrix and zero field splitting D-tensor
-
-        #Rotate cartesian tensors to crystal coordinates
-        self.g_tensor = self.rot_mat @ g_tensor @ self.rot_mat.T 
-        self.d_tensor = self.rot_mat @ d_tensor @ self.rot_mat.T
-
-        hours_input, minutes_input, seconds_input = self.timer(timer_input)
-
-                        # Matrix initialization
-        self.Hs = np.zeros([self.hdim, self.hdim], dtype=np.complex128)
-        self.eigenvalues = np.zeros(self.hdim, dtype=np.complex128)
[…]
