### 2025-10-07 96363a6
Delete redfield.py
