+++ redfield.py
@@ -1,354 +1,0 @@
-import numpy as np
-from constants import hbar, c, hbar_SI, cm2J
-import math_func
-import coupling
-from mpi4py import MPI
-from matplotlib import pyplot as plt
-from tqdm import tqdm
-
-class Redfield:
-
-   def __init__(self, S, T, eigenvectors,eigenvalues,q_vector, omega_q, Delta_alpha_q, L_vectors):
-
-      self.S = S
-      self.T = T
-      self.Ns = int(2*self.S + 1)
-      self.hdim = self.Ns
-      self.eigenvectors = eigenvectors # Eigenvectors of the spin Hamiltonian
-      self.eigenvalues = eigenvalues
-      self.q_vector = q_vector # Phonon mode index and wave vector q 
-      self.omega_q = omega_q
-      self.Delta_alpha_q = Delta_alpha_q # Broadening parameter
-      self.Nq = L_vectors.shape[0] # Number of q points
-      self.Nomega = L_vectors.shape[1] # Number of phonons at q
-
-      self.omega = self.eigenvalues[:, None] - self.eigenvalues[None, :]
-
-   def G_1ph(self,omega_ij, omega_alpha_q):
-      """
-
-      Parameters:
-         omega_ij (float): Energy difference (ω_ij).
-         omega_alpha_q (float): Phonon frequency (ω_{αq}).
-         Delta_alpha_q (float): Broadening parameter (Δ_{αq}).
-         n_alpha_q (float): Phonon occupation number (n̄_{αq}).
-      
-      Returns:
-         float: Value of G^{1-ph}.
-      """
-
-      #cm-1 to rad/s conversion
-      omega_alpha_q = omega_alpha_q 
-      omega_ij =  omega_ij 
-      Delta_alpha_q = self.Delta_alpha_q 
-
-      delta = math_func.lorentzian((omega_ij - omega_alpha_q), Delta_alpha_q)
-      n_aq = math_func.bose_einstein(omega_alpha_q, self.T)
-      G_1ph = delta * n_aq + delta * (n_aq + 1)
-
-      #term1_numerator = Delta_alpha_q
-      #term1_denominator = Delta_alpha_q**2 + (omega_ij - omega_alpha_q)**2
-      #n_1 = math_func.bose_einstein(omega_alpha_q, self.T)
-      #term1 = (term1_numerator / term1_denominator) * n_1
-      
-      #term2_numerator = Delta_alpha_q
-      #term2_denominator = Delta_alpha_q**2 + (omega_ij + omega_alpha_q)**2
-      #n_2 = math_func.bose_einstein(omega_alpha_q, self.T)
-      #term2 = (term2_numerator / term2_denominator) * (n_2 + 1)
-      
-      #G_1ph = (term1 + term2) 
-      G_1ph = G_1ph / (2* np.pi* c)           
-      return G_1ph
-
-   def G_2ph(self, omega_ij, omega_alpha_q, omega_beta_qp):
-
-      #cm-1 to rad/s conversion
-      omega_alpha_q = omega_alpha_q
-      omega_beta_qp = omega_beta_qp 
-      omega_ij = omega_ij 
-      Delta = self.Delta_alpha_q
-
-      Delta = self.Delta_alpha_q + self.Delta_alpha_q
-      
-      n_alpha = math_func.bose_einstein(omega_alpha_q, self.T)
-      n_beta = math_func.bose_einstein(omega_beta_qp, self.T)
-
-      # Four terms of the Green's function
-      term1 = math_func.lorentzian(omega_ij - omega_alpha_q - omega_beta_qp, Delta) * n_alpha * n_beta
[…]
