### 2025-10-07 22b418f
Delete measure.py
