### 2025-10-07 fe6ce58
Delete math_func.py
