+++ math_func.py
@@ -1,194 +1,0 @@
-import numpy as np
-from constants import k_B, kB_SI
-from numpy.polynomial.polynomial import polyfit, polyval2d, Polynomial
-
-def lorentzian(x, eta):
-
-    return (eta / np.pi) / ( (x) ** 2 + eta ** 2)
-
-def diagonalize(hamiltonian):
-    """
-    Diagonalize a Hamiltonian.
-    """
-
-    eigenvalues, eigenvectors = np.linalg.eigh(hamiltonian)
-    return eigenvalues, eigenvectors
-
-
-def energy_diff(eigenvalues):
-    """
-    Compute the energy difference.
-    """
-
-    energy_diff = eigenvalues[:, None] - eigenvalues[None, :]
-    return energy_diff
-
-
-def mat(a, b, alpha,q,eigenvectors, V_alpha):
-
-    tmp = np.dot(V_alpha[q,alpha,:, :],eigenvectors[b])
-    tmp1 = np.dot(eigenvectors[a], tmp)
-
-    return tmp1
-
-def bose_einstein(omega_alpha_q, T):
-        """
-        Compute the Bose-Einstein occupation number.
-        
-        Parameters:
-            omega_alpha_q (float): Phonon frequency (ω_{αq}).
-            T (float): Temperature in Kelvin.
-        
-        Returns:
-            float: Bose-Einstein occupation number (n̄_{αq}).
-        """
-        x = omega_alpha_q / (k_B * T)
-        x = np.clip(x, None, 700)
-        
-        n_alpha_q = 1 / (np.exp(x) - 1)
-
-        return n_alpha_q
-
-def finite_difference(f_plus, f_minus, delta):
-    """
-    Computes the numerical gradient at 0 using central difference.
-
-    Parameters:
-        f_plus  : float or np.array
-            Function value at +delta
-        f_zero  : float or np.array
-            Function value at 0
-        f_minus : float or np.array
-            Function value at -delta
-        delta   : float
-            Displacement step size
-
-    Returns:
-        float or np.array: Estimated derivative at 0
-    """
-    return (f_plus - f_minus) / (2 * delta)
-
-def compute_derivative(x,fx,displacement=0.0,degree=3):
-    """
-    Compute the derivative of f_x with reespect to x by polynomial fitting
-
-    Returns: df_x / dx (array): Derivative of f_x with respect to x at 0.
-
-    """
-    
-    coefficients = np.polyfit(x, fx, degree)
-    
-    # Calculate derivative coefficients (using polyder)
-    deriv_coeffs = np.polyder(coefficients)
-    
-    # Evaluate derivative at displacement
-    dfdx = np.polyval(deriv_coeffs, displacement)
-
-    return dfdx
-
-
-def compute_second_derivative(x, fx, poly_degree=3):
-    """
-    Fit a 2D polynomial to the function values and compute the mixed second derivative at (0, 0).
-    
-    Parameters:
-        x: np.ndarray of shape (2, N) — x[0] contains x values, x[1] contains y values.
-        fx: np.ndarray of shape (M, M) — function values on the x-y grid.
-        poly_degree: int — degree of the 2D polynomial.
-
-    Returns:
-        float: second derivative ∂²f/∂x∂y at (0, 0).
-    """
-    x_vals, y_vals = x[0], x[1]
-    
-    # Handle potential dimension mismatch by using the actual fx dimensions
-    if fx.shape[0] != len(x_vals) or fx.shape[1] != len(y_vals):
-        # Trim or extend the coordinate arrays to match fx dimensions
[…]
