### 2025-10-07 44c27d8
Delete hamiltonian.py
