### 2025-10-07 1e8596c
Delete coupling.py
