+++ coupling.py
@@ -1,273 +1,0 @@
-import numpy as np
-from constants import hbar_SI, c
-import hamiltonian
-import math_func
-from mpi4py import MPI
-
-class coupling:
-    def __init__(self, B, S, T, eigenvectors, q_vector, omega_q, R_vectors, L_vectors,rot_mat,Ncells, file_reader):
-    
-        self.file_reader = file_reader
-        self.B = B # Magnetic field in cm-1
-        self.S = S # Spin
-        self.T = T # Temperatur in K
-        self.Ns = int(2*self.S + 1) # Number of spins
-        self.hdim = self.Ns
-        self.eigenvectors = eigenvectors # Eigenvectors of the spin Hamiltonian
-        self.indices = self.file_reader.read_indices() # Mapping of the molecule indices in the crystal
-        self.N = len(self.indices) # Number of atoms in the molecule
-        self.masses = self.file_reader.read_mol_masses() # Atomic mass of the molecule
-        self.q_vector = q_vector # Phonon mode index and wave vector q 
-        self.omega_q = omega_q # Phonon frequencies in cm-1
-        self.Nq = L_vectors.shape[0] # Number of q points
-        self.Nomega = L_vectors.shape[1] # Number of phonons at q
-        self.Ncells = Ncells
-        self.L_vectors = L_vectors # Phonon eigenvectors
-        self.R_vectors = R_vectors # Atomic position vectors
-        self.R_vectors_mol = R_vectors[self.indices,:] # Position vectors of the molecule in the crystal
-        self.L_vectors_mol = np.zeros((self.Nq, self.Nomega, self.N, 3), dtype=np.complex128) # Eigenvectors of the molecule in the crystal
-        self.L_vectors_mol = L_vectors[:,:,self.indices,:] 
-        self.N_atoms = R_vectors.shape[0]  # Number of atoms in the crystal
-        self.rot_mat = rot_mat # Rotational matrix for hte molecule to match the coordinates in the crystal
-        
-        
-        d_tensor, g_tensor = self.file_reader.read_orca() # g-matrix and ZFS D-tensor
-
-        #Rotate the molecular cartesian matrices to match the crystal coordinates
-        self.g_tensor = self.rot_mat @ g_tensor @ self.rot_mat.T
-        self.d_tensor = self.rot_mat @ d_tensor @ self.rot_mat.T
-
-        self.dH_dx = np.zeros((self.N,3,self.hdim,self.hdim), dtype=np.complex128)
-
-        #Read linear displacement g-matrix and ZFS D-tensor
-        D_d1, G_d1,self.disp1 = self.file_reader.read_d1()
-        self.G_d1 = self.rot_mat @ G_d1 @ self.rot_mat
-        self.D_d1 = self.rot_mat @ D_d1 @ self.rot_mat
-        self.compute_dH_dx() #Compute dH/dx
-
-        #Read quadratic displacement g-matrix and ZFS D-tensor
-        D_d2, G_d2, self.disp2 = self.file_reader.read_d2()
-        self.G_d2 = self.rot_mat @ G_d2 @ self.rot_mat
-        self.D_d2 = self.rot_mat @ D_d2 @ self.rot_mat
-
-        self.dH2_dxdx = np.zeros((self.N,self.N,3,3,self.hdim,self.hdim), dtype=np.complex128)
-        self.compute_d2H_dxdx() #Compute d2H/dxdx'
-
-        #Compute linear spin-phonon coupling
-        self.V_alpha = np.zeros((self.Nq, self.Nomega,self.hdim, self.hdim), dtype=np.complex128)
-        self.V_alpha = self.compute_V_alpha_q()
-
[…]
