### 2025-10-07 3c15a92
Delete constants.py
