+++ constants.py
@@ -1,15 +1,0 @@
-"""Constants"""
-
-eV2cm=8065.61042 #cm-1
-bohr2ang = 0.5291772108 #angstrom
-cm2J = 1.98644568E-23 # J/cm-1
-c = 2.99792458E10 #cm/s
-#hbar = 6.582119569E-16 #eV*s
-hbar = 6.582119569E-16 * eV2cm
-hbar_SI = 1.0545718E-34 # J*s
-#k_B = 8.617333262E-5 #eV/K
-kB_SI = 1.38064852E-23 # J/K
-k_B = 8.617333262E-5 * eV2cm #cm-1/K
-avogadro = 6.02214179E23 #mol-1
-#Bohrmagneton = 5.7883818060E-5 #eV/T
-Bohrmagneton = 0.46686 #cm-1/T
