+++ docs/index.md
@@ -1,0 +1,25 @@
+# MyProject
+
+Welcome to the **MyProject** documentation!
+
+MyProject is a Python library for predicting material properties using machine learning models.  
+It combines Density Functional Theory (DFT) data with Graph Neural Networks (GNNs) for accurate property prediction.
+
+---
+
+## 🚀 Features
+
+- 📦 Easy-to-use API for training and inference  
+- 🧠 Support for GNN, e3nn, and transformer-based architectures  
+- ⚙️ Automatic preprocessing of `.mol`, `.xyz`, and `.cif` files  
+- 🧪 Integration with Quantum ESPRESSO and VASP workflows  
+- 📊 Visualization utilities for band structures and tensors  
+
+---
+
+## 📥 Installation
+
+You can install MyProject directly from GitHub:
+
+```bash
+pip install git+https://github.com/yourusername/myproject.git
