+++ docs/index.md
@@ -1,25 +1,28 @@
-# MyProject
+# SpinVibe
 
-Welcome to the **MyProject** documentation!
 
-MyProject is a Python library for predicting material properties using machine learning models.  
-It combines Density Functional Theory (DFT) data with Graph Neural Networks (GNNs) for accurate property prediction.
-
----
-
-## 🚀 Features
-
-- 📦 Easy-to-use API for training and inference  
-- 🧠 Support for GNN, e3nn, and transformer-based architectures  
-- ⚙️ Automatic preprocessing of `.mol`, `.xyz`, and `.cif` files  
-- 🧪 Integration with Quantum ESPRESSO and VASP workflows  
-- 📊 Visualization utilities for band structures and tensors  
+**SpinVibe** is an open-source Python package for simulating **spin–phonon coupling in molecular qubits** using first-principles calculations.  
+It is desgin to study molecular qubits in solid-state systems.
 
 ---
 
-## 📥 Installation
+## Inputs
+- Solid-state phonons and eigenvectors
+- Molecular spin-hamiltonian parameters (e.g. g-factor, zero-field splitting tensor)
+- Parameters: Temperature, Polarization, external magnetic field, etc.
 
-You can install MyProject directly from GitHub:
+## Output
+- Time evolution of the spin density and magnetization
+- Spin-phonon relaxation time (T1)
 
-```bash
-pip install git+https://github.com/yourusername/myproject.git
+---
+
+## Requirements
+The code was written using the following Python Libraries:
+
+- python                    3.9.21
+- numpy                     1.26.4
+- scipy                     1.13.1
+- mpi4py                    4.0.2
+- h5py                      3.12.1
+- tqdm                      4.67.1 (optional for tracking parallerization performance)
