### 2025-10-07 835ddda
Update redfield.py
