+++ src/redfield.py
@@ -4,7 +4,6 @@ import math_func
 import coupling
 from mpi4py import MPI
 from matplotlib import pyplot as plt
-from tqdm import tqdm
 
 class Redfield:
 
@@ -194,10 +193,6 @@ class Redfield:
 
       R2_local = np.zeros((self.hdim, self.hdim, self.hdim, self.hdim), dtype=np.complex128)
 
-      # Setup progress bar only for rank 0
-      #if rank == 0:
-      #   pbar = tqdm(total=len(all_tasks), desc="Computing R2 tensor", unit="tasks", ncols=100)
-
       for q1, q2, alpha, beta in all_tasks:
          # Get V matrix for this (alpha, beta) pair
          V_alpha_beta = init_Vq.compute_V_alpha_beta_q(q1, q2, alpha, beta)
@@ -252,103 +247,9 @@ class Redfield:
                         
                         # Apply prefactor and add to result
                         R2_local[a, b, c, d] += prefactor * term_sum
-            # Update progress bar only for rank 0
-        # if rank == 0:
-        #       pbar.update(1)
-
-      # Close progress bar for rank 0
-      #if rank == 0:
-      #   pbar.close()
 
       # Gather results from all processes
       R2_tensor = np.zeros((self.hdim, self.hdim, self.hdim, self.hdim), dtype=np.complex128)
       comm.Allreduce(R2_local, R2_tensor, op=MPI.SUM)
       
       return R2_tensor
-
-   def R4_tensor(self, V_alpha):
-      comm = MPI.COMM_WORLD
-      rank = comm.Get_rank()
-      size = comm.Get_size()
-
-      hdim = V_alpha.shape[2]
-      Nq = V_alpha.shape[0]
-      Nomega = V_alpha.shape[1]
-
-      all_tasks = []
-      for q in range(self.Nq):
-         for alpha in range(self.Nomega):
-               all_tasks.append((q, alpha))
-
-      all_tasks = np.array_split(all_tasks, size)[rank]
-
-      # Precompute matrix elements M[q, alpha]
-      M = np.zeros((Nq, Nomega, hdim, hdim), dtype=np.complex128)
-      for q, alpha in all_tasks:
-         V = V_alpha[q, alpha]
-         M[q, alpha] = self.eigenvectors.conj().T @ V @ self.eigenvectors
-
-      # Energy differences and tensor initialization
-      omega_ij = math_func.energy_diff(self.eigenvalues)
-      R4_local = np.zeros((hdim, hdim, hdim, hdim), dtype=np.complex128)
-
-      for q, alpha in all_tasks:
-         for beta in range(Nomega):
-            if alpha >= beta:
-               continue
-            delta_alpha_beta = 1 if alpha == beta else 0
-
-            M_alpha = M[q, alpha]
-            M_beta = M[q, beta]
-
-            omega_q_alpha = self.omega_q[q, alpha]
-            omega_q_beta = self.omega_q[q, beta]
-
-            # Compute all W terms
-            term1 = M_alpha @ (M_beta.T / (omega_ij - omega_q_beta))
-            term2 = M_beta @ (M_alpha.T / (omega_ij - omega_q_alpha))
-            W_mm = np.abs(term1 + term2)**2
-
-            term3 = M_alpha @ (M_beta.T / (omega_ij + omega_q_beta))
-            term4 = M_beta @ (M_alpha.T / (omega_ij + omega_q_alpha))
-            W_pp = np.abs(term3 + term4)**2
-
-            term5 = M_alpha @ (M_beta.T / (omega_ij + omega_q_beta))
[…]
