+++ README.md
@@ -1,28 +1,96 @@
-# SpinVibe
+<style>
+.container {
+  display: flex;
+  gap: 2rem;
+  max-width: 1200px;
+}
 
+.sidebar {
+  flex: 0 0 200px;
+  position: sticky;
+  top: 20px;
+  height: fit-content;
+}
 
-**SpinVibe** is an open-source Python package for simulating **spin–phonon coupling in molecular qubits** using first-principles calculations.  
-It is desgin to study molecular qubits in solid-state systems.
+.sidebar nav {
+  background: #f5f5f5;
+  padding: 1rem;
+  border-radius: 8px;
+  border: 1px solid #ddd;
+}
+
+.sidebar h3 {
+  margin-top: 0;
+  font-size: 1rem;
+  margin-bottom: 0.75rem;
+}
+
+.sidebar ul {
+  list-style: none;
+  padding: 0;
+  margin: 0;
+}
+
+.sidebar li {
+  margin: 0.5rem 0;
+}
+
+.sidebar a {
+  text-decoration: none;
+  color: #333;
+}
+
+.sidebar a:hover {
+  color: #0066cc;
+}
+
+.content {
+  flex: 1;
+  min-width: 0;
+}
+</style>
+
+<div class="container">
+  <aside class="sidebar">
+    <nav>
+      <h3>Contents</h3>
+      <ul>
+        <li><a href="#step-1">1. Download Source</a></li>
+        <li><a href="#step-2">2. Navigate Directory</a></li>
+        <li><a href="#step-3">3. Install Code</a></li>
+      </ul>
+    </nav>
+  </aside>
+
+  <main class="content">
+
+## SpinVibe Installation Guide
+
+### <span id="step-1">Step 1: Download Source Code</span>
+
+Download SpinVibe source code using the command:
+
+```bash
+git clone https://github.com/victorchanglee/SpinVibe.git
+```
 
 ---
 
-## Inputs
-- Solid-state phonons and eigenvectors
-- Molecular spin-hamiltonian parameters (e.g. g-factor, zero-field splitting tensor)
-- Parameters: Temperature, Polarization, external magnetic field, etc.
+### <span id="step-2">Step 2: Navigate to Root Directory</span>
 
-## Output
-- Time evolution of the spin density and magnetization
-- Spin-phonon relaxation time (T1)
+```bash
+cd SpinVibe
+```
 
 ---
 
-## Requirements
-The code was written using the following Python Libraries:
+### <span id="step-3">Step 3: Install the Code</span>
 
-- python                    3.9.21
-- numpy                     1.26.4
-- scipy                     1.13.1
-- mpi4py                    4.0.2
-- h5py                      3.12.1
-- tqdm                      4.67.1 (optional for tracking parallerization performance)
+Install the code with the command:
+
+```bash
+pip install -e .
+```
+
+  </main>
+</div>
