+++ README.md
@@ -1,72 +1,34 @@
-<style>
-.container {
-  display: flex;
-  gap: 2rem;
-  max-width: 1200px;
-}
+# SpinVibe
 
-.sidebar {
-  flex: 0 0 200px;
-  position: sticky;
-  top: 20px;
-  height: fit-content;
-}
 
-.sidebar nav {
-  background: #f5f5f5;
-  padding: 1rem;
-  border-radius: 8px;
-  border: 1px solid #ddd;
-}
+**SpinVibe** is an open-source Python package for simulating **spin–phonon coupling in molecular qubits** using first-principles calculations.  
+It is desgin to study molecular qubits in solid-state systems.
 
-.sidebar h3 {
-  margin-top: 0;
-  font-size: 1rem;
-  margin-bottom: 0.75rem;
-}
+---
 
-.sidebar ul {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
+## Inputs
+- Solid-state phonons and eigenvectors
+- Molecular spin-hamiltonian parameters (e.g. g-factor, zero-field splitting tensor)
+- Parameters: Temperature, Polarization, external magnetic field, etc.
 
-.sidebar li {
-  margin: 0.5rem 0;
-}
+## Output
+- Time evolution of the spin density and magnetization
+- Spin-phonon relaxation time (T1)
 
-.sidebar a {
-  text-decoration: none;
-  color: #333;
-}
+---
 
-.sidebar a:hover {
-  color: #0066cc;
-}
+## Requirements
+The code was written using the following Python Libraries:
 
-.content {
-  flex: 1;
-  min-width: 0;
-}
-</style>
-
-<div class="container">
-  <aside class="sidebar">
-    <nav>
-      <h3>Contents</h3>
-      <ul>
-        <li><a href="#step-1">1. Download Source</a></li>
-        <li><a href="#step-2">2. Navigate Directory</a></li>
-        <li><a href="#step-3">3. Install Code</a></li>
-      </ul>
-    </nav>
-  </aside>
-
-  <main class="content">
+- python                    3.9.21
+- numpy                     1.26.4
+- scipy                     1.13.1
+- mpi4py                    4.0.2
+- h5py                      3.12.1
 
 ## SpinVibe Installation Guide
 
-### <span id="step-1">Step 1: Download Source Code</span>
+### Step 1: Download Source Code
 
 Download SpinVibe source code using the command:
 
@@ -76,7 +38,7 @@ git clone https://github.com/victorchanglee/SpinVibe.git
 
 ---
 
-### <span id="step-2">Step 2: Navigate to Root Directory</span>
+### Step 2: Navigate to Root Directory
 
 ```bash
 cd SpinVibe
@@ -84,13 +46,10 @@ cd SpinVibe
 
 ---
 
-### <span id="step-3">Step 3: Install the Code</span>
+### Step 3: Install the Code
 
 Install the code with the command:
 
 ```bash
 pip install -e .
 ```
-
-  </main>
-</div>
