+++ docs/index.md
@@ -1,57 +1,4 @@
-<style>
-.container {
-  display: flex;
-  gap: 2rem;
-  max-width: 1200px;
-  margin: 0 auto;
-}
 
-.sidebar {
-  flex: 0 0 220px;
-  position: sticky;
-  top: 20px;
-  height: fit-content;
-}
-
-.sidebar nav {
-  background: #f5f5f5;
-  padding: 1.25rem;
-  border-radius: 8px;
-  border: 1px solid #ddd;
-}
-
-.sidebar h3 {
-  margin-top: 0;
-  font-size: 1rem;
-  margin-bottom: 1rem;
-  color: #333;
-}
-
-.sidebar ul {
-  list-style: none;
-  padding: 0;
-  margin: 0;
-}
-
-.sidebar li {
-  margin: 0.5rem 0;
-}
-
-.sidebar a {
-  text-decoration: none;
-  color: #555;
-  font-size: 0.9rem;
-}
-
-.sidebar a:hover {
-  color: #0066cc;
-}
-
-.content {
-  flex: 1;
-  min-width: 0;
-}
-</style>
 
 <div class="container">
   <aside class="sidebar">
