+++ docs/index.md
@@ -1,56 +1,34 @@
+# SpinVibe
 
-
-<div class="container">
-  <aside class="sidebar">
-    <nav>
-      <h3>Table of Contents</h3>
-      <ul>
-        <li><a href="#overview">Overview</a></li>
-        <li><a href="#inputs">Inputs</a></li>
-        <li><a href="#output">Output</a></li>
-        <li><a href="#requirements">Requirements</a></li>
-        <li><a href="#installation">Installation Guide</a></li>
-        <li style="margin-left: 1rem;"><a href="#step-1">Step 1: Download</a></li>
-        <li style="margin-left: 1rem;"><a href="#step-2">Step 2: Navigate</a></li>
-        <li style="margin-left: 1rem;"><a href="#step-3">Step 3: Install</a></li>
-      </ul>
-    </nav>
-  </aside>
-
-  <main class="content">
-
-# <span id="overview">SpinVibe</span>
 
 **SpinVibe** is an open-source Python package for simulating **spin–phonon coupling in molecular qubits** using first-principles calculations.  
-It is designed to study molecular qubits in solid-state systems.
+It is desgin to study molecular qubits in solid-state systems.
 
 ---
 
-## <span id="inputs">Inputs</span>
-
+## Inputs
 - Solid-state phonons and eigenvectors
 - Molecular spin-hamiltonian parameters (e.g. g-factor, zero-field splitting tensor)
 - Parameters: Temperature, Polarization, external magnetic field, etc.
 
-## <span id="output">Output</span>
-
+## Output
 - Time evolution of the spin density and magnetization
 - Spin-phonon relaxation time (T1)
 
 ---
 
-## <span id="requirements">Requirements</span>
-
+## Requirements
 The code was written using the following Python Libraries:
+
 - python                    3.9.21
 - numpy                     1.26.4
 - scipy                     1.13.1
 - mpi4py                    4.0.2
 - h5py                      3.12.1
 
-## <span id="installation">SpinVibe Installation Guide</span>
+## SpinVibe Installation Guide
 
-### <span id="step-1">Step 1: Download Source Code</span>
+### Step 1: Download Source Code
 
 Download SpinVibe source code using the command:
 
@@ -60,7 +38,7 @@ git clone https://github.com/victorchanglee/SpinVibe.git
 
 ---
 
-### <span id="step-2">Step 2: Navigate to Root Directory</span>
+### Step 2: Navigate to Root Directory
 
 ```bash
 cd SpinVibe
@@ -68,13 +46,10 @@ cd SpinVibe
 
 ---
 
-### <span id="step-3">Step 3: Install the Code</span>
+### Step 3: Install the Code
 
 Install the code with the command:
 
 ```bash
 pip install -e .
 ```
-
-  </main>
-</div>
