+++ docs/index.md
@@ -1,55 +1,380 @@
-# SpinVibe
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>SpinVibe Documentation</title>
+    <style>
+        * {
+            margin: 0;
+            padding: 0;
+            box-sizing: border-box;
+        }
 
+        body {
+            font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif;
+            line-height: 1.6;
+            color: #404040;
+        }
 
-**SpinVibe** is an open-source Python package for simulating **spin–phonon coupling in molecular qubits** using first-principles calculations.  
-It is desgin to study molecular qubits in solid-state systems.
+        /* Top Navigation Bar */
+        .top-nav {
+            background: #2980b9;
+            color: white;
+            padding: 0;
+            position: fixed;
+            top: 0;
+            left: 0;
+            right: 0;
+            z-index: 1000;
+            height: 50px;
+            display: flex;
+            align-items: center;
+            box-shadow: 0 2px 4px rgba(0,0,0,0.1);
+        }
 
----
+        .nav-content {
+            display: flex;
+            align-items: center;
+            justify-content: space-between;
+            width: 100%;
+            padding: 0 20px;
+        }
 
-## Inputs
-- Solid-state phonons and eigenvectors
-- Molecular spin-hamiltonian parameters (e.g. g-factor, zero-field splitting tensor)
-- Parameters: Temperature, Polarization, external magnetic field, etc.
+        .nav-left {
+            display: flex;
+            align-items: center;
+            gap: 20px;
+        }
 
-## Output
-- Time evolution of the spin density and magnetization
-- Spin-phonon relaxation time (T1)
+        .project-name {
+            font-size: 18px;
+            font-weight: 600;
+            color: white;
+            text-decoration: none;
+        }
 
----
+        .version-selector {
+            background: rgba(255,255,255,0.2);
+            border: none;
+            color: white;
+            padding: 5px 10px;
+            border-radius: 3px;
+            font-size: 14px;
+            cursor: pointer;
+        }
 
-## Requirements
-The code was written using the following Python Libraries:
+        .search-box {
+            background: rgba(255,255,255,0.9);
+            border: none;
+            padding: 6px 12px;
+            border-radius: 3px;
+            width: 250px;
+            font-size: 14px;
+        }
 
-- python                    3.9.21
-- numpy                     1.26.4
-- scipy                     1.13.1
-- mpi4py                    4.0.2
-- h5py                      3.12.1
+        /* Main Container */
+        .container {
+            display: flex;
+            margin-top: 50px;
+            min-height: calc(100vh - 50px);
+        }
 
-## SpinVibe Installation Guide
+        /* Sidebar */
+        .sidebar {
+            width: 300px;
+            background: #f9f9f9;
+            border-right: 1px solid #e1e4e5;
+            position: fixed;
+            left: 0;
+            top: 50px;
+            bottom: 0;
+            overflow-y: auto;
+            padding: 20px 0;
+        }
 
-### Step 1: Download Source Code
+        .sidebar-content {
+            padding: 0 20px;
+        }
 
-Download SpinVibe source code using the command:
+        .sidebar h3 {
+            color: #2980b9;
+            font-size: 16px;
+            margin: 20px 0 10px 0;
+            font-weight: 600;
+        }
 
-```bash
-git clone https://github.com/victorchanglee/SpinVibe.git
-```
+        .sidebar ul {
+            list-style: none;
+            margin: 0 0 15px 0;
+        }
 
----
+        .sidebar li {
+            margin: 5px 0;
+        }
 
[…]
