+++ docs/index.md
@@ -1,276 +1,51 @@
 
+<!DOCTYPE html>
 <html lang="en">
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>SpinVibe Documentation</title>
-    <style>
-        * {
-            margin: 0;
-            padding: 0;
-            box-sizing: border-box;
-        }
-
-        body {
-            font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "Helvetica Neue", Arial, sans-serif;
-            line-height: 1.6;
-            color: #404040;
-        }
-
-
-        /* Main Container */
-        .container {
-            display: flex;
-            margin-top: 50px;
-            min-height: calc(100vh - 50px);
-        }
-
-        /* Sidebar */
-        .sidebar {
-            width: 300px;
-            background: #f9f9f9;
-            border-right: 1px solid #e1e4e5;
-            position: fixed;
-            left: 0;
-            top: 50px;
-            bottom: 0;
-            overflow-y: auto;
-            padding: 20px 0;
-        }
-
-        .sidebar-content {
-            padding: 0 20px;
-        }
-
-        .sidebar h3 {
-            color: #2980b9;
-            font-size: 16px;
-            margin: 20px 0 10px 0;
-            font-weight: 600;
-        }
-        .sidebar ul {
-            list-style: none;
-            margin: 0 0 15px 0;
-        }
-
-
-        .sidebar li {
-            margin: 5px 0;
-        }
-
-        .sidebar a {
-            color: #404040;
-            text-decoration: none;
-            display: block;
-            padding: 5px 10px;
-            border-radius: 3px;
-            font-size: 14px;
-            transition: background 0.2s;
-        }
-
-        .sidebar a:hover {
-            background: #e1e4e5;
-        }
-
-        .sidebar .sub-item {
-            margin-left: 15px;
-            font-size: 13px;
-        }
-
-        /* Main Content */
-        .main-content {
-            margin-left: 300px;
-            padding: 40px 60px;
-            max-width: 1000px;
-            flex: 1;
-        }
-
-        .main-content h1 {
-            color: #2c3e50;
-            font-size: 2.5em;
-            margin-bottom: 20px;
-            font-weight: 600;
-        }
-
-        .main-content h2 {
-            color: #2c3e50;
-            font-size: 1.8em;
-            margin: 40px 0 20px 0;
-            padding-bottom: 10px;
-            border-bottom: 2px solid #e1e4e5;
-            font-weight: 600;
-        }
-
-        .main-content h3 {
-            color: #2c3e50;
-            font-size: 1.3em;
-            margin: 30px 0 15px 0;
-            font-weight: 600;
-        }
-
-        .main-content p {
-            margin-bottom: 15px;
-            line-height: 1.7;
-        }
-
-        .main-content ul {
-            margin: 15px 0 15px 30px;
-        }
-
-        .main-content li {
-            margin: 8px 0;
-        }
-
-        .main-content code {
-            background: #f5f5f5;
-            padding: 2px 6px;
-            border-radius: 3px;
-            font-family: 'Courier New', monospace;
-            font-size: 0.9em;
-            color: #e74c3c;
-        }
-
-        .main-content pre {
-            background: #f8f8f8;
-            border: 1px solid #e1e4e5;
-            border-left: 3px solid #2980b9;
-            padding: 15px;
-            border-radius: 3px;
-            overflow-x: auto;
-            margin: 20px 0;
-        }
-
-        .main-content pre code {
-            background: none;
-            padding: 0;
-            color: #333;
-        }
-
-        .note {
-            background: #e7f2fa;
[…]
