### 2025-10-07 e711e7d
Update cli.py
