### 2025-10-07 e46c915
Update setup.py
