### 2025-10-07 #9154 @victorchanglee
@kyleniemeyer thank you so much for looking into my submission. I'm currently addressing all the submission issues.

So far I uploaded an OSI-approved license, installation instructions, input files for testing and I think the code is structured to be installable using pip.

I should be able to finish with the documentation in the next couple of days.

Please let me know if they are comply with the guidelines or I need to amend something.
