+++ README.md
@@ -1,5 +1,4 @@
-HTML: <a href="https://joss.theoj.org/papers/6cd884ec29554707741cd8e700542a68"><img src="https://joss.theoj.org/papers/6cd884ec29554707741cd8e700542a68/status.svg"></a>
-Markdown: [![status](https://joss.theoj.org/papers/6cd884ec29554707741cd8e700542a68/status.svg)](https://joss.theoj.org/papers/6cd884ec29554707741cd8e700542a68)
+[![status](https://joss.theoj.org/papers/6cd884ec29554707741cd8e700542a68/status.svg)](https://joss.theoj.org/papers/6cd884ec29554707741cd8e700542a68)
 [![Documentation](https://victorchanglee.github.io/SpinVibe/)](https://victorchanglee.github.io/SpinVibe/)
 
 # SpinVibe
