### 2025-10-07 12be5bc
Update coupling.py
