### 2025-10-07 bf1f61e
Update math_func.py
