### 2025-10-07 fb0f874
Update redfield.py
