### 2025-10-07 32aa57e
Update coupling.py
