### 2025-10-07 fc64b7e
Update redfield.py
