+++ src/redfield.py
@@ -1,5 +1,5 @@
 import numpy as np
-from constants import hbar, c, hbar_SI, cm2J
+from .constants import hbar, c, hbar_SI, cm2J
 from . import math_func
 from . import coupling
 from mpi4py import MPI
