### 2025-10-07 c0dcbdc
Update spin_phonon.py
