### 2025-10-07 295db85
Update coupling.py
