### 2025-10-10 e45f18b
Update and rename setup.py to pyproject.toml
