### 2025-10-10 dd255b6
Update and rename pyproject.toml to setup.py
