### 2025-10-10 39e5cc5
Delete setup.py
