+++ src/README.md
@@ -1,1 +1,0 @@
-# spin-phonon
