### 2025-10-10 824c536
Update coupling.py
