+++ test/README.md
@@ -1,1 +1,6 @@
+To run this test, simply run
+
+```bash
+spinvibe input.json
+```
 
