+++ README.md
@@ -68,4 +68,4 @@ python -m build
 pip install dist/spinvibe-1.0-py3-none-any.whl
 ```
 
-
+Once the installation is succesfull, you can test it using the files provided in the test directory.
