### 2025-10-14 9db58c7
Update coupling.py
