+++ spinvibe/redfield.py
@@ -71,29 +71,23 @@ class Redfield:
       n_alpha = math_func.bose_einstein(omega_alpha_q, self.T)
       n_beta = math_func.bose_einstein(omega_beta_qp, self.T)
 
-      # Four terms of the Green's function
       term1 = math_func.lorentzian(omega_ij - omega_alpha_q - omega_beta_qp, Delta) * n_alpha * n_beta
       term2 = math_func.lorentzian(omega_ij + omega_alpha_q + omega_beta_qp, Delta) * (n_alpha + 1) * (n_beta + 1)
       term3 = math_func.lorentzian(omega_ij + omega_alpha_q - omega_beta_qp, Delta) * (n_alpha + 1) * n_beta
       term4 = math_func.lorentzian(omega_ij - omega_alpha_q + omega_beta_qp, Delta) * n_alpha * (n_beta + 1)
-
-      #term1 = (Delta / (Delta**2 + (omega_ij - omega_alpha_q - omega_beta_qp)**2)) * n_alpha * n_beta
-      #term2 = (Delta / (Delta**2 + (omega_ij + omega_alpha_q + omega_beta_qp)**2)) * (n_alpha + 1) * (n_beta + 1)
-      #term3 = (Delta / (Delta**2 + (omega_ij - omega_alpha_q + omega_beta_qp)**2)) * n_alpha * (n_beta + 1)
-      #term4 = (Delta / (Delta**2 + (omega_ij + omega_alpha_q - omega_beta_qp)**2)) * (n_alpha + 1) * n_beta
 
       G2ph = (term1 + term2 + term3 + term4) / (2*np.pi*c)
 
 
       return G2ph
 
-   def R1_tensor(self, V_alpha):
+   def R1_tensor(self, init_Vq):
       comm = MPI.COMM_WORLD
       rank = comm.Get_rank()
       size = comm.Get_size()
 
       omega_alpha = self.omega_q
-      V_alpha = V_alpha * cm2J
+      
 
       prefactor = -np.pi / (2 * hbar_SI**2)
       
@@ -113,7 +107,8 @@ class Redfield:
       for q, alpha in tasks:
          # Initialize contribution for this (q, alpha) pair
          R1_qa = np.zeros((self.hdim, self.hdim, self.hdim, self.hdim), dtype=np.complex128)
-         
+         V_alpha = init_Vq.compute_V_alpha_q(q, alpha)
+         V_alpha = V_alpha * cm2J
          # Loop over tensor indices
          for a in range(self.hdim):
                for b in range(self.hdim):
@@ -124,15 +119,13 @@ class Redfield:
                            term1 = 0.0
                            if b == d:
                               for j in range(self.hdim):
-                                 V_aj = V_alpha[q, alpha, a, j]
-                                 V_jc = V_alpha[q, alpha, j, c]
                                  G_term1 = self.G_1ph(self.omega[j, c], omega_alpha[q, alpha])
-                                 term1 += V_aj * V_jc * G_term1
+                                 term1 += V_alpha[a, j] * V_alpha[j, c] * G_term1
                         
                            #print(term1, G_term1)
                            # Term 2: V^α_{ac} V^α_{db} G^{1-ph}(ω_{bd}, ω_α)
-                           V_ac = V_alpha[q, alpha, a, c]
-                           V_db = V_alpha[q, alpha, d, b]
+                           V_ac = V_alpha[a, c]
+                           V_db = V_alpha[d, b]
                            G_term2 = self.G_1ph(self.omega[b, d], omega_alpha[q, alpha])
                            term2 = V_ac * V_db * G_term2
 
@@ -144,10 +137,8 @@ class Redfield:
                            term4 = 0.0
                            if c == a:
[…]
