### 2025-10-21 1114f51
Remove parallelization setup for (q, omega) pairs
Removed code for creating and distributing (q, omega) pairs.
