+++ spinvibe/constants.py
@@ -3,7 +3,7 @@
 eV2cm=8065.61042 #cm-1
 bohr2ang = 0.5291772108 #angstrom
 cm2J = 1.98644568E-23 # J/cm-1
-c = 2.99792458E10 #cm/s
+c_cms = 2.99792458E10 #cm/s
 hbar = 6.582119569E-16 * eV2cm
 hbar_SI = 1.0545718E-34 # J*s
 kB_SI = 1.38064852E-23 # J/K
