### 2025-12-03 #9154 @RMeli
Thank you @MartinBeseda!

> Also, I'm curious - is it possible to ask for a confirmation of one's review on OrcID?

You can follow the discussion and current work about this here: https://github.com/openjournals/joss/issues/813.
