### 2025-12-15 victorchanglee/SpinVibe#1 @apizzuto
Paper nitpicks
Part of [JOSS Review](https://github.com/openjournals/joss-reviews/issues/9538):

I'm going through the paper and had some very tiny little pedantic nitpicks that I thought I'd just drop in an issue. Overall, the paper is well written and meets the criteria of a JOSS submission. I will begin my review of the code shortly.

  * Figure 1 caption: typo (environmental)
  * It looks like there is some mis-rendered tex in the pdf, namely there are visible $ and similarly some math text that is not rendered in math mode, and there are also a few % following or preceding equations.
  * The phrase "To describe the time evolution of the system, we employ the secular Redfield equation under the the Born-Markov approximation " is repeated (and also has two "the"-s next to each other)
