### 2025-12-15 victorchanglee/SpinVibe#2 @apizzuto
Installation issues
Beginning to do the installation for [https://github.com/openjournals/joss-reviews/issues/9538](https://github.com/openjournals/joss-reviews/issues/9538).

On the documentation: it seems like both of the options in your README are "3-a". Is this intended?

Also, I had some problems with installing the package. I think this comes from the fact that mpi4py requires a working MPI implementation, so I had to do that before being able to run the code without hitting MPI related errors. The below is what I got when I tried using the documentation as written.

```
error: Cannot compile MPI programs. Check your configuration!!!
      Installing mpi4py requires a working MPI implementation.
      If you are running on a supercomputer or cluster, check with
      the system administrator or refer to the system user guide.
      Otherwise, if you are running on a laptop or desktop computer,
      your may be missing the MPICH or Open MPI development package:
      * On Fedora/RHEL systems, run:
        $ sudo dnf install mpich-devel     # for MPICH
        $ sudo dnf install openmpi-devel   # for Open MPI
      * On Debian/Ubuntu systems, run:
        $ sudo apt install libmpich-dev    # for MPICH
        $ sudo apt install libopenmpi-dev  # for Open MPI
```

It might be worth calling out that the users will need to do the MPI installation first?

One other thing on the installation: I think that pip uses pyproject.toml by default, so the different steps in the installation (the different 3-a steps) confused me. I could be wrong, but this might be worth double checking (I'm suspicious that this is the case because for the MPI issues I tried deleting the dependencies in the setup.py file and running `pip install -e .`, but it still saw the dependencies, and it didn't stop trying to install the dependencies untill I removed them from the pyproject.toml file and THEN ran `pip install -e .`)
