### 2025-12-15 victorchanglee/SpinVibe#3 @apizzuto
Code Issues
Part of [JOSS Review](https://github.com/openjournals/joss-reviews/issues/9538):

I ran the test as suggested and got the following output:

```
Reading input parameters from: input.json
Initializing input files...
  spin_file: Cr_otolyl_4.h5
  phonon_file: phonon.h5
  d1_file: Cr_otolyl_4_d1.h5
  d2_file: d_16.h5
  g2_file: g_16.h5
  atoms_file: Sn_otolyl_4_atoms.h5
  indices_file: molecule_indices.h5
  mol_mass: mol_mass.h5
2025-12-15 17:19:50
---------------------------------------
|Start Spin-phonon coupling simulation|
---------------------------------------

Input Parameters
================
Magnetic field: [0.   0.   0.41]
S: 1
T: 20.0
Broadening: 1.0
Population type: pure
Polarization: [0 0 1]
Rotational matrix:
[[0 0 1]
 [1 0 0]
 [0 1 0]]

Spin Hamiltonian
[[-3.61376628e-01+0.00000000e+00j -2.91812008e-06+1.06333248e-05j
   5.11553213e-02-7.05205722e-06j]
 [-2.91812008e-06-1.06333248e-05j -3.41022461e-02+0.00000000e+00j
   2.91812008e-06-1.06333248e-05j]
 [ 5.11553213e-02+7.05205722e-06j  2.91812008e-06+1.06333248e-05j
   3.95478874e-01+0.00000000e+00j]]

Eigenvalues of the spin Hamiltonian
[-0.36481853 -0.03410225  0.39892077]

Initialize simulation
=====================
Running in serial mode.

Computing linear and quadratic coupling
Initializing R1 tensor (Linear coupling)
Number of tasks per process: 342
Eigenvalues of the R1 matrix
[-4.32928403e+05-1.33043085e-11j -3.73096574e+05-1.50990218e-11j
 -2.51849298e+05+1.14899361e-11j -2.08380892e+05+4.80470640e-12j
 -1.75882555e-10-1.13631736e-12j -3.67404673e+04+1.14670298e-11j
 -5.45373231e+04-1.39158160e-11j -8.53366127e+04-7.04282517e-12j
 -9.73847207e+04+8.18470136e-12j]

Initializing R2 tensor (Quadratic coupling)
Number of tasks per process: 58653
Eigenvalues of the R2 matrix
[-2.08534136e+08+1.00621537e-08j -2.00838125e+08+3.26564506e-09j
 -1.87097051e+08+3.45102388e-10j -1.84273465e+08+1.56270768e-10j
 -1.40143977e+07-8.61465807e-09j -1.23104329e+07+3.15338789e-10j
  8.29796548e-07+5.26322023e-09j -3.21323375e+06-4.07869328e-09j
 -5.74226878e+06+8.18678157e-09j]

Eigenvalues of the Redfield matrix
[-2.08897837e+08-6.60129009e-09j -2.01136734e+08+2.31104090e-09j
 -1.87383028e+08-8.71619105e-09j -1.84493332e+08+8.21804180e-09j
  8.09357876e-07-1.69890279e-09j -1.41058892e+07+2.43914779e-09j
 -1.24205488e+07-8.95466710e-09j -3.25784591e+06-8.88792559e-09j
 -5.86814915e+06-8.26082222e-09j]

Initial spin population:
[[0.+0.j 0.+0.j 0.+0.j]
 [0.+0.j 0.+0.j 0.+0.j]
 [0.+0.j 0.+0.j 1.+0.j]]

Start time evolution

/usr/local/google/home/apizzuto/SpinVibe/spinvibe/spin_phonon.py:342: RuntimeWarning: overflow encountered in matmul
  k2 = self.R_mat @ (rho[i] + 0.5 * self.dt * k1)
/usr/local/google/home/apizzuto/SpinVibe/spinvibe/spin_phonon.py:342: RuntimeWarning: invalid value encountered in matmul
  k2 = self.R_mat @ (rho[i] + 0.5 * self.dt * k1)
/usr/local/google/home/apizzuto/SpinVibe/spinvibe/spin_phonon.py:343: RuntimeWarning: invalid value encountered in multiply
  k3 = self.R_mat @ (rho[i] + 0.5 * self.dt * k2)
/usr/local/google/home/apizzuto/SpinVibe/spinvibe/spin_phonon.py:346: RuntimeWarning: overflow encountered in multiply
  rho[i+1] = rho[i] + (self.dt / 6) * (k1 + 2*k2 + 2*k3 + k4)
/usr/local/google/home/apizzuto/SpinVibe/spinvibe/spin_phonon.py:346: RuntimeWarning: invalid value encountered in multiply
  rho[i+1] = rho[i] + (self.dt / 6) * (k1 + 2*k2 + 2*k3 + k4)
/usr/local/google/home/apizzuto/spinvibe/lib/python3.13/site-packages/numpy/linalg/_linalg.py:2828: RuntimeWarning: overflow encountered in multiply
  s = (x.conj() * x).real
/usr/local/google/home/apizzuto/SpinVibe/spinvibe/measure.py:89: RuntimeWarning: invalid value encountered in scalar subtract
  if abs(self.Mz[i] - self.Mz[i - 1]) < tolerance:
Equilibrium M_parallel: nan
Fit failed: array must not contain infs or NaNs
T1 from magnetization decay
T1 =  None s
T1_err =  None s
Traceback (most recent call last):
  File "/usr/local/google/home/apizzuto/spinvibe/bin/spinvibe", line 7, in <module>
    sys.exit(main())
             ~~~~^^
  File "/usr/local/google/home/apizzuto/SpinVibe/spinvibe/run.py", line 51, in main
    spin_phonon.spin_phonon(
    ~~~~~~~~~~~~~~~~~~~~~~~^
        B=B,
        ^^^^
    ...<10 lines>...
        R_type=params["R_type"]
        ^^^^^^^^^^^^^^^^^^^^^^^
    )
    ^
  File "/usr/local/google/home/apizzuto/SpinVibe/spinvibe/spin_phonon.py", line 233, in __init__
    if self.T1 < self.T1_err:
       ^^^^^^^^^^^^^^^^^^^^^
TypeError: '<' not supported between instances of 'NoneType' and 'NoneType'
```

As noted in the installation issue I opened, I had some problems installing the package with the default package versions, so I ended up having the following versions. I'm not sure if that is what is causing the issue or if there is a problem with the code. Any input would be valued.

```
h5py     3.15.1
mpi4py   4.0.2
numpy    2.3.5
pip      25.3
scipy    1.16.3
spinvibe 1.0     /usr/local/google/home/apizzuto/SpinVibe
```
