### 2025-12-15 victorchanglee/SpinVibe#4 @apizzuto
Example documentation request
Hello! As part of [JOSS Review](https://github.com/openjournals/joss-reviews/issues/9538), I had a request. The code right now seems like a bit of a black box, and normally for JOSS submissions it is nice to have an example / tutorial that explains what the code is doing instead of just pointing to input and saying that the code on the input produces the desired output.

Is there any way you could add something like a jupyter notebook or a README in addition to the tests? It could have the exact same technical contents as the test file, but it would be nice to walk through the example in a bit more detail, describing what's going on in a few more steps than just "run this command and note that the output is what we want" (because to a new user, they won't understand why the output is correct).

I don't consider this a blocking issue for acceptance into JOSS, but I think it would be nice to have.
