### 2025-12-15 victorchanglee/SpinVibe#5 @apizzuto
Contribution Guide
Part of [JOSS Review](https://github.com/openjournals/joss-reviews/issues/9538):

One of the requirements of a JOSS package is the following:

Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

I didn't notice this anywhere in the repo. Apologies if I missed it, but if you do not yet have this, could you add this?
