### 2025-12-15 #9538 @apizzuto
Hello authors! I've done an initial pass of the review. Overall I found the paper compelling and I think the code is documented well. I'm having some problems running the code and I have some documentation related requests that I've opened linked issues about. Looking forward to hearing back.
