### 2026-01-06 3dc0144
Fix typo in environmental factors description
