### 2026-01-06 22d1568
Update paper.md
