### 2026-01-06 213bb22
Update paper.md
