+++ docs/paper.md
@@ -107,15 +107,15 @@ weak-coupling regime the interaction between the spin Hamiltonian and
 the phonons is expanded as a Taylor series truncated at the quadratic
 term:
 
-\begin{equation}
+```math
 \hat{H}_{sph} = \sum_{\alpha q} \left( \frac{\partial \hat{H}_s}{\partial Q_{\alpha q}} \right) 
 \hat{Q}_{\alpha q}(t) 
 + \sum_{\alpha q} \sum_{\beta q'} 
 \left( \frac{\partial^2 \hat{H}_s}{\partial Q_{\alpha q} \partial Q_{\beta q'}} \right)
 \hat{Q}_{\alpha q}(t) \hat{Q}_{\beta q'}(t)\,. 
 \label{eq:1}
-\end{equation}
-%
+```
+
 In \autoref{eq:1}, the first order term represents the interaction of the spin with a single phonon $ Q_{\alpha q}$. The second order term describes the interaction of the spin with two phonons $ Q_{\alpha q}$ and $Q_{\beta q'}$. 
 
 To describe the time evolution of the system, we employ the secular Redfield equation under the Born-Markov approximation 
