### 2026-01-06 d6dc1b9
Update paper.md
