### 2026-01-06 8275bc5
Update paper.md
