### 2026-01-06 2a10e13
Fix Hamiltonian equation formatting in paper.md
Corrected formatting of the Hamiltonian equation in the paper.
