### 2026-01-06 d7360d0
Add equation for weak-coupling regime in paper
