### 2026-01-06 d740250
Update paper.md
