### 2026-01-06 0ab1c24
Update paper.md
