### 2026-01-06 2f6a58e
Include equation format in paper.md
Added equation representation in the documentation.
