+++ docs/paper.md
@@ -99,6 +99,8 @@ HDF5 files, it is not restricted to specific codes. Any ab-initio
 package can be used to generate the input data, provided that the
 relevant quantities are extracted and saved in this format.
 
+$$ a = x $$
+
 # First principles spin-phonon coupling 
 
 Our code implements the theoretical framework developed by Lunghi and
