### 2026-01-06 8f7eced
Include spin-phonon coupling equations in paper
Added equations for spin-phonon coupling in the paper.
