### 2026-01-06 33adf7e
Correct formatting of Hamiltonian equation in paper.md
