### 2026-01-06 569380d
Insert equation for variables in paper documentation
Add mathematical expression for variables 'a' and 'b'.
