### 2026-01-06 531e995
Update paper.md
