### 2026-01-06 0f297f2
Update equation in paper.md
