### 2026-01-06 c730aa5
Clean up LaTeX equations in paper.md
Removed unnecessary LaTeX math block and adjusted equations.
