### 2026-01-06 ac2a968
Revise Redfield superoperator equations in paper
Updated equations related to the Redfield superoperator and phonon contributions, including corrections to the formatting and mathematical expressions.
