+++ docs/paper.md
@@ -135,7 +135,7 @@ $$
 
 where $R_{ab,cd}$ denotes the Redfield superoperator. The single-phonon contribution to the Redfield tensor, under the secular approximation, is
 
-
+$$
 R^{1\text{-ph}}_{ab,cd} = -\frac{\pi}{2 \hbar^2} 
 \sum_{\alpha} \Bigg\{ 
 &\sum_{j} \delta_{bd} V^{\alpha}_{aj} V^{\alpha}_{jc} 
@@ -146,22 +146,23 @@ G^{1\text{-ph}}(\omega_{ac}, \omega_{\alpha}) \\
 G^{1\text{-ph}}(\omega_{bd}, \omega_{\alpha}) 
 + \sum_{j} \delta_{ca} V^{\alpha}_{dj} V^{\alpha}_{jb} 
 G^{1\text{-ph}}(\omega_{jd}, \omega_{\alpha}) 
-\Bigg\}\,,
-
+\Bigg\}
+$$
 
 where $V^{\alpha} = \left( \frac{\partial \hat{H}_s}{\partial Q_{\alpha}} \right)$ and 
 
-
+$$
 G^{1\text{--ph}}(\omega_{ij}, \omega_{\alpha q})
 = \frac{1}{\pi} \left[
 \frac{\Delta_{\alpha q}}{\Delta_{\alpha q}^{2} + (\omega_{ij} - \omega_{\alpha q})^{2}} \, \bar{n}_{\alpha q}
 + \frac{\Delta_{\alpha q}}{\Delta_{\alpha q}^{2} + (\omega_{ij} + \omega_{\alpha q})^{2}} \, (\bar{n}_{\alpha q}+1)
 \right]
+$$
 
 is the phonon correlation function in the presence of anharmonic terms that leads to a decay in the phonon lifetime with a rate $\Delta_{\alpha q}$. 
 
 The two-phonon contribution is given by
-
+$$
 R^{2\text{--ph}}_{ab,cd} 
 = -\frac{\pi}{4 \hbar^2} \sum_{\alpha \geq \beta } \Bigg\{ \,
 & \sum_j  \delta_{bd} V^{\alpha \beta}_{aj} V^{\alpha \, \beta }_{jc} 
@@ -173,9 +174,11 @@ R^{2\text{--ph}}_{ab,cd}
 +& \sum_j  \delta_{ca} V^{\alpha \beta}_{dj} V^{\alpha \, \beta}_{jb} 
    G^{2\text{--ph}}(\omega_{jd}, \omega_{\alpha}, \omega_{\beta}) 
 \Bigg\}
+$$
 
 where $V^{\alpha \beta} = \left( \frac{\partial^2 \hat{H}_s}{\partial Q_{\alpha} \partial Q_{\beta}} \right)$ and 
 
+$$
 G^{2\text{--ph}}(\omega_{ij},\omega_{\alpha q},\omega_{\beta q'}) 
 = \frac{1}{\pi} \Bigg[ \,
 &\frac{\Delta_{\alpha q \beta q'}}
@@ -191,19 +194,19 @@ G^{2\text{--ph}}(\omega_{ij},\omega_{\alpha q},\omega_{\beta q'})
        {\Delta_{\alpha q \beta q'}^{2} + (\omega_{ij} + \omega_{\alpha q} - \omega_{\beta q'})^{2}}
        (\bar{n}_{\alpha q}+1)\bar{n}_{\beta q'} 
 \Bigg]
-
+$$
 
 Once the time-evolution is performed, we can study the time-evolution of the magnetization of the system 
 
-
-    \vec{M}(t) = \text{Tr} (\vec{S} \hat{\rho} (t))
-
+$$
[…]
