### 2026-01-06 48158d7
Fix formatting of Redfield tensor equation in paper.md
Corrected the formatting and structure of the Redfield tensor equation in the document.
