### 2026-01-06 62f2619
Refactor single-phonon Redfield tensor equation
Refactor the single-phonon contribution to the Redfield tensor for clarity and consistency.
