### 2026-01-06 b8dd0b5
Update paper.md
