### 2026-01-06 f811103
Update paper.md
