+++ docs/paper.md
@@ -139,7 +139,7 @@ $$
 R^{1\text{-ph}}_{ab,cd}
 = -\frac{\pi}{2\hbar^{2}}
 \sum_{\alpha}
-\left\{
+{
 &\sum_{j} \delta_{bd}\,
 V^{\alpha}_{aj} V^{\alpha}_{jc}\,
 G^{1\text{-ph}}(\omega_{jc}, \omega_{\alpha}) \\
@@ -150,8 +150,7 @@ G^{1\text{-ph}}(\omega_{bd}, \omega_{\alpha}) \\
 &+ \sum_{j} \delta_{ca}\,
 V^{\alpha}_{dj} V^{\alpha}_{jb}\,
 G^{1\text{-ph}}(\omega_{jd}, \omega_{\alpha})
-\right\{
-
+}
 $$
 
 where $V^{\alpha} = \left( \frac{\partial \hat{H}_s}{\partial Q_{\alpha}} \right)$ and 
