### 2026-01-06 254612d
Simplify equation for R^{1-ph}_{ab,cd} in paper
Removed unnecessary terms and simplified the equation for R^{1-ph}_{ab,cd}.
