### 2026-01-06 61beac1
Update paper.md
