### 2026-01-06 1b8aa9c
Update paper.md
