### 2026-01-06 77c3e25
Update paper.md
