### 2026-01-06 72d9a39
Fix LaTeX syntax in Redfield tensor equation
Correct LaTeX formatting for the Redfield tensor equation.
