### 2026-01-06 058c8fc
Refactor phonon correlation function equation
Removed redundant equation for the phonon correlation function and clarified its significance in the context of anharmonic terms.
