### 2026-01-06 4b45aa0
Fix equation formatting in paper.md
Corrected formatting and removed unnecessary symbols in the equation for the phonon correlation function.
