### 2026-01-06 f000c7b
Fix equation formatting in docs/paper.md
Correct the formatting of the phonon correlation function equation in the paper.
