### 2026-01-06 c511f56
Refactor two-phonon contribution equation
Reformatted the two-phonon contribution equation for clarity.
