### 2026-01-06 f656256
Add two-phonon contribution equation to paper
