### 2026-01-06 c37816d
Refactor G^{2--ph} expression for clarity
Simplified the expression for G^{2--ph} in the paper.
