+++ docs/paper.md
@@ -156,21 +156,7 @@ $$
 where $V^{\alpha \beta} = \left( \frac{\partial^2 \hat{H}_s}{\partial Q_{\alpha} \partial Q_{\beta}} \right)$ and 
 
 $$
-G^{2\text{--ph}}(\omega_{ij},\omega_{\alpha q},\omega_{\beta q'}) 
-= \frac{1}{\pi} \Bigg[ \,
-&\frac{\Delta_{\alpha q \beta q'}}
-       {\Delta_{\alpha q \beta q'}^{2} + (\omega_{ij} - \omega_{\alpha q} - \omega_{\beta q'})^{2}}
-       \bar{n}_{\alpha q}\bar{n}_{\beta q'} \nonumber \\[6pt]
-+&\frac{\Delta_{\alpha q \beta q'}}
-       {\Delta_{\alpha q \beta q'}^{2} + (\omega_{ij} + \omega_{\alpha q} + \omega_{\beta q'})^{2}}
-       (\bar{n}_{\alpha q}+1)(\bar{n}_{\beta q'}+1) \nonumber \\[6pt]
-+&\frac{\Delta_{\alpha q \beta q'}}
-       {\Delta_{\alpha q \beta q'}^{2} + (\omega_{ij} - \omega_{\alpha q} + \omega_{\beta q'})^{2}}
-       \bar{n}_{\alpha q}(\bar{n}_{\beta q'}+1) \nonumber \\[6pt]
-+&\frac{\Delta_{\alpha q \beta q'}}
-       {\Delta_{\alpha q \beta q'}^{2} + (\omega_{ij} + \omega_{\alpha q} - \omega_{\beta q'})^{2}}
-       (\bar{n}_{\alpha q}+1)\bar{n}_{\beta q'} 
-\Bigg]
+G^{2\text{--ph}}(\omega_{ij},\omega_{\alpha q},\omega_{\beta q'}) = \frac{1}{\pi} [ \frac{\Delta_{\alpha q \beta q'}}{\Delta_{\alpha q \beta q'}^{2} + (\omega_{ij} - \omega_{\alpha q} - \omega_{\beta q'})^{2}} \bar{n}_{\alpha q}\bar{n}_{\beta q'} + \frac{\Delta_{\alpha q \beta q'}} {\Delta_{\alpha q \beta q'}^{2} + (\omega_{ij} + \omega_{\alpha q} + \omega_{\beta q'})^{2}} (\bar{n}_{\alpha q}+1)(\bar{n}_{\beta q'}+1) + \frac{\Delta_{\alpha q \beta q'}} {\Delta_{\alpha q \beta q'}^{2} + (\omega_{ij} - \omega_{\alpha q} + \omega_{\beta q'})^{2}} \bar{n}_{\alpha q}(\bar{n}_{\beta q'}+1) + \frac{\Delta_{\alpha q \beta q'}} {\Delta_{\alpha q \beta q'}^{2} + (\omega_{ij} + \omega_{\alpha q} - \omega_{\beta q'})^{2}} (\bar{n}_{\alpha q}+1)\bar{n}_{\beta q'} ]
 $$
 
 Once the time-evolution is performed, we can study the time-evolution of the magnetization of the system 
