### 2026-01-06 f26726f
Fix formatting in equations and derivatives section
