### 2026-01-06 2dbc286
Update paper.md
