### 2026-01-06 fd15fa2
Fix LaTeX formatting in paper.md
