### 2026-01-06 1b9a3da
Fix notation for Hamiltonian derivative in equation
Corrected formatting of the Hamiltonian derivative notation in the equation.
