### 2026-01-06 1518b1c
Fix formatting and punctuation in paper.md
Corrected formatting and punctuation in the explanation of phonon interactions and the Redfield equation.
