### 2026-01-06 321b9d3
Update paper.md
