### 2026-01-06 7b182ea
Clarify derivative notation in paper.md
