### 2026-01-06 c6bb4dd
Update paper.md
