### 2026-01-06 a6801b1
Update paper.md
