### 2026-01-06 c57a34b
Update paper.md
