### 2026-01-06 b48ac5c
Fix spacing in equation for V^{\alpha}
Corrected spacing in the equation for V^{\alpha}.
