### 2026-01-06 f6099ec
Correct equation formatting and notation in paper.md
Fixed formatting and added missing partial derivative notation in the equation.
