### 2026-01-06 ef9bb15
Simplify equation by removing duplicate definition
Removed redundant definition of V^{\alpha} in the equation.
