### 2026-01-06 5976af3
Update paper.md
