### 2026-01-06 a3d34ab
Fix Hamiltonian notation in paper.md
Corrected the notation for the Hamiltonian in the equation.
