### 2026-01-06 ffca7b3
Fix LaTeX formatting in paper.md
