+++ docs/paper.md
@@ -153,7 +153,7 @@ $$
 R^{2\text{--ph}}_{ab,cd} = -\frac{\pi}{4 \hbar^2} \sum_{\alpha \geq \beta } \\{ \sum_j  \delta_{bd} V^{\alpha \beta}_{aj} V^{\alpha \, \beta }_{jc} G^{2\text{--ph}}(\omega_{jc}, \omega_{\alpha }, \omega_{\beta}) -V^{\alpha \beta }_{ac} V^{\alpha \, \beta }_{db}  G^{2\text{--ph}}(\omega_{ac}, \omega_{\alpha}, \omega_{\beta}) - V^{\alpha \beta }_{ac} V^{\alpha \, \beta }_{db} G^{2\text{--ph}}(\omega_{bd}, \omega_{\alpha }, \omega_{\beta }) + \sum_j  \delta_{ca} V^{\alpha \beta}_{dj} V^{\alpha \, \beta}_{jb} G^{2\text{--ph}}(\omega_{jd}, \omega_{\alpha}, \omega_{\beta}) \\}
 $$
 
-where $V^{\alpha \beta} = \frac{\partial^2 \hat{H}_{s} {\partial Q_{\alpha} \partial Q_{\beta}} $ and 
+where $V^{\alpha \beta} = \frac{\partial^2 \hat{H_{s}}}{\partial Q_{\alpha} \partial Q_{\beta}} $ and 
 
 $$
 G^{2\text{--ph}}(\omega_{ij},\omega_{\alpha q},\omega_{\beta q'}) = \frac{1}{\pi} [ \frac{\Delta_{\alpha q \beta q'}}{\Delta_{\alpha q \beta q'}^{2} + (\omega_{ij} - \omega_{\alpha q} - \omega_{\beta q'})^{2}} \bar{n}_{\alpha q}\bar{n}_{\beta q'} + \frac{\Delta_{\alpha q \beta q'}} {\Delta_{\alpha q \beta q'}^{2} + (\omega_{ij} + \omega_{\alpha q} + \omega_{\beta q'})^{2}} (\bar{n}_{\alpha q}+1)(\bar{n}_{\beta q'}+1) + \frac{\Delta_{\alpha q \beta q'}} {\Delta_{\alpha q \beta q'}^{2} + (\omega_{ij} - \omega_{\alpha q} + \omega_{\beta q'})^{2}} \bar{n}_{\alpha q}(\bar{n}_{\beta q'}+1) + \frac{\Delta_{\alpha q \beta q'}} {\Delta_{\alpha q \beta q'}^{2} + (\omega_{ij} + \omega_{\alpha q} - \omega_{\beta q'})^{2}} (\bar{n}_{\alpha q}+1)\bar{n}_{\beta q'} ]
