### 2026-01-06 e5f65e0
Update paper.md
