### 2026-01-06 d935daa
Update paper.md
