### 2026-01-14 victorchanglee/SpinVibe#6 @MartinBeseda
Comparison to S-o-A
Part of [JOSS Review](https://github.com/openjournals/joss-reviews/issues/9538):

The code is not compared to some state-of-the-art frameworks, mainly [spin_phonon_suite](https://pypi.org/project/spin-phonon-suite/).
I believe it's important to pinpoint main differences here to see, what's new/better in SpinVibe.
