### 2026-01-14 victorchanglee/SpinVibe#7 @MartinBeseda
Missing release
Part of [JOSS Review](https://github.com/openjournals/joss-reviews/issues/9538):

There's is not release or even tag denoting the version being published on the repository. It'd be better to "fix" the stable version like this.
