### 2026-01-14 #9538 @MartinBeseda
@victorchanglee @RMeli I've created a few more issues and put some notes on some already-existing ones. I'm still testing the installation process and the functionality, will put the that notes here today or tomorrow at last.
