### 2026-01-15 818fdaa
Update paper.md
