+++ test/README.md
@@ -1,7 +1,0 @@
-To run this test, simply run
-
-```bash
-spinvibe input.json
-```
-
-The expected output for a serial run is in test.out
