+++ setup.py
@@ -13,9 +13,11 @@ setup(
     install_requires=[
         "numpy==1.26.4",
         "scipy==1.13.1",
-        "mpi4py==4.0.2",
         "h5py==3.12.1",
     ],
+    optional_require={
+        "mpi": ["mpi4py==4.0.2"],
+    },
     entry_points={
         "console_scripts": [
             "spinvibe = spinvibe.run:main",
