### 2026-01-16 victorchanglee/SpinVibe#2 @victorchanglee
@apizzuto Thanks for the comments. Indeed, mpi4py requires a working MPI implementation. I updated the code to make mpi4py optional and thus it can be now run in serial and clarify the installation steps.
