+++ spinvibe/spin_phonon.py
@@ -15,7 +15,7 @@ except ImportError:
     MPI = None
 
 class spin_phonon:
-    def __init__(self, B, S, Delta_alpha_q, rot_mat, pol, T, tf, dt, file_reader,save_file,init_type='polarized',R_type=None):
+    def __init__(self, B, S, Delta_alpha_q, rot_mat, pol, T, tf, dt, file_reader,save_file,init_type='polarized'):
         
         # Get MPI info if available, otherwise use serial mode
         if mpi_on:
@@ -46,7 +46,6 @@ class spin_phonon:
         self.pol = pol  # Polarization vector
         self.T = T  # Temperature in Kelvin
         self.init_type = init_type
-        self.R_type = R_type
         self.save_file = save_file
         if rank == 0:
             print("Input Parameters")
@@ -120,14 +119,12 @@ class spin_phonon:
             else:
                 print("Running in serial mode.")
                 print("\n")
-        
 
-            if self.R_type == None:
-                print("Computing linear and quadratic coupling")
-            if self.R_type == 'R1':
-                print("Computing ONLY linear coupling")
-            if self.R_type == 'R2':
-                print("Computing ONLY quadratic coupling")
+        # Initialize spin density
+
+        self.init_occ = np.zeros(self.hdim, dtype=np.complex128)
+        self.rho0 = np.zeros([self.hdim**2], dtype=np.complex128)
+        self.rho0 = self.init_rho()
 
         init_Vq = coupling.coupling(self.B, self.S, self.T, self.eigenvectors,self.q_vector, self.omega_q, self.R_vectors, self.L_vectors,self.rot_mat,self.file_reader,self.save_file)
 
@@ -142,48 +139,96 @@ class spin_phonon:
         self.R1 = np.zeros((self.hdim, self.hdim, self.hdim, self.hdim), dtype=np.complex128)
         self.R2 = np.zeros((self.hdim, self.hdim, self.hdim, self.hdim), dtype=np.complex128)
 
-        if self.R_type in (None, 'R1'):
+        if rank == 0:
+            print("Initializing R1 tensor (Linear coupling)")
 
-            if rank == 0:
-                print("Initializing R1 tensor (Linear coupling)")
+        timer_R1 = time.perf_counter()
+        self.R1 = self.init_R.R1_tensor(init_Vq)
+        hours_R1, minutes_R1, seconds_R1 = self.timer(timer_R1)
 
-            timer_R1 = time.perf_counter()
-            self.R1 = self.init_R.R1_tensor(init_Vq)
-            hours_R1, minutes_R1, seconds_R1 = self.timer(timer_R1)
+        if rank == 0:
             R1_mat = self.R1.reshape((self.hdim**2, self.hdim**2)) #Transform into matrix form
 
             eigenvalues, eigenvectors = np.linalg.eig(R1_mat)
+            print("Eigenvalues of the R1 matrix")
+            print(eigenvalues)
+            print("\n")
+            
+            self.tf = tf  # Total time
+            self.dt = dt  # Time step
+            self.tlist = np.linspace(0, self.tf, int(self.tf / self.dt))
+            self.tsteps = len(self.tlist)
 
-            if rank == 0:
-                print("Eigenvalues of the R1 matrix")
-                print(eigenvalues)
-                print("\n")
-                
+            self.drho_dt = self.time_evolution(eigenvalues, eigenvectors)
+
+            self.rho_t = np.zeros([self.hdim, self.hdim, self.tsteps], dtype=np.complex128)
+
+            self.rho_t = self.drho_dt.reshape(self.tsteps, self.hdim, self.hdim).transpose(2, 1, 0)
+
+
[…]
