### 2026-01-29 2cba39c
Update paper.md
