### 2026-01-29 3e871f6
Update paper.md
