### 2026-01-29 victorchanglee/SpinVibe#6 @victorchanglee
@MartinBeseda We updated the paper.md to address existing codes that are design to solve the same scientific questions. While the codes work in the same theoretical framework, we want to emphasize that our code is design with a user flexibility and ease of use in mind.
