+++ spinvibe/coupling.py
@@ -13,7 +13,7 @@ except ImportError:
     MPI = None
 
 class coupling:
-    def __init__(self, B, S, T, eigenvectors, q_vector, omega_q, R_vectors, L_vectors,rot_mat, file_reader,save_file):
+    def __init__(self, B, S, T, eigenvectors, q_vector, omega_q, R_vectors, L_vectors,masses, rot_mat, disp1, disp2, file_reader,save_file):
 
         # Get MPI info if available, otherwise use serial mode
         if mpi_on:
@@ -34,16 +34,20 @@ class coupling:
         self.eigenvectors = eigenvectors # Eigenvectors of the spin Hamiltonian
         self.indices = self.file_reader.read_indices() # Mapping of the molecule indices in the crystal
         self.N = len(self.indices) # Number of atoms in the molecule
-        self.masses = self.file_reader.read_mol_masses() # Atomic mass of the molecule
+        self.masses = masses[self.indices] # Atomic mass of the molecule
+
         self.q_vector = q_vector # Phonon mode index and wave vector q 
         self.omega_q = omega_q # Phonon frequencies in cm-1
         self.Nq = L_vectors.shape[0] # Number of q points
         self.Nomega = L_vectors.shape[1] # Number of phonons at q 
-        self.L_vectors = L_vectors # Phonon eigenvectors
-        self.R_vectors = R_vectors # Lattice vectors
+
+        self.R_vectors_mol = np.zeros((self.N, 3), dtype=np.complex128)
+        self.R_vectors_mol = R_vectors[self.indices,:] # Lattice vectors
         self.L_vectors_mol = np.zeros((self.Nq, self.Nomega, self.N, 3), dtype=np.complex128) # Eigenvectors of the molecule in the crystal
         self.L_vectors_mol = L_vectors[:,:,self.indices,:] 
         self.rot_mat = rot_mat # Rotational matrix for hte molecule to match the coordinates in the crystal
+        self.disp1 = disp1
+        self.disp2 = np.stack([disp2,disp2], axis=0) # Displacements for finite difference
         
         d_tensor, g_tensor = self.file_reader.read_spin() # g-matrix and ZFS D-tensor
 
@@ -51,18 +55,16 @@ class coupling:
         self.g_tensor = self.rot_mat @ g_tensor @ self.rot_mat.T
         self.d_tensor = self.rot_mat @ d_tensor @ self.rot_mat.T
 
-        self.dH_dx = np.zeros((self.N,3,self.hdim,self.hdim), dtype=np.complex128)
-
         #Read linear displacement g-matrix and ZFS D-tensor
-        D_d1, G_d1,self.disp1 = self.file_reader.read_d1()
+        D_d1, G_d1, D_d2, G_d2 = self.file_reader.read_derivatives()
+
         self.G_d1 = self.rot_mat @ G_d1 @ self.rot_mat
         self.D_d1 = self.rot_mat @ D_d1 @ self.rot_mat
-        self.compute_dH_dx() #Compute dH/dx
-
-        #Read quadratic displacement g-matrix and ZFS D-tensor
-        D_d2, G_d2, self.disp2 = self.file_reader.read_d2()
         self.G_d2 = self.rot_mat @ G_d2 @ self.rot_mat
         self.D_d2 = self.rot_mat @ D_d2 @ self.rot_mat
+
+        self.dH_dx = np.zeros((self.N,3,self.hdim,self.hdim), dtype=np.complex128)
+        self.compute_dH_dx() #Compute dH/dx
 
         self.dH2_dxdx = np.zeros((self.N,self.N,3,3,self.hdim,self.hdim), dtype=np.complex128)
[…]
