### 2026-02-09 cf2e751
Delete tools/extract_d1.py
