### 2026-02-17 #9538 @MartinBeseda
@victorchanglee Thank you for all the changes, I don't see any big problem with your repo now. Only minor nitpick - in the Contribution and Support section you're a) talking about Pull and Fork workflow, while you probably mean Pull Request and Fork and b) you're not explicitly stating any contact for questions. It might be cool to do that either as a special issue category alongside Feature Requests and Bugs or provide some contact person. But this is only a minor recommendation, the version is ready to be published as it is, in my opinion.

@RMeli Thank you for the collaboration here. In my opinion, the paper is ready to be published.
