### 2026-02-18 b13df32
Update contact section wording in README
