+++ spinvibe/measure.py
@@ -52,84 +52,74 @@ class measure:
 
         return
     
-    
     def calc_t1(self):
         """
-        Compute the T1 time from the Mz(t) signal using exponential fitting.
-        Returns T1 and its uncertainty.
+        Compute T1 using logarithmic transformation for linear fitting.
+        Cuts data when Mz drops below 1e-2.
         """
-        
-        # Get MPI rank if available, otherwise use rank 0 for serial mode
         if mpi_on:
             comm = MPI.COMM_WORLD
             rank = comm.Get_rank()
         else:
             rank = 0
 
-        # polarization axis
+        # Get Mz data
         n = np.array(self.pol if self.pol is not None else [0., 0., 1.], dtype=float)
-        n_norm = np.linalg.norm(n)
-        if n_norm == 0:
-            raise ValueError("self.Mpol must not be the zero vector")
-        n /= n_norm
-
+        n /= np.linalg.norm(n)
         
-        # Extract Mz(t) data and time points
-
         if self.init_type == 'polarized':
-            if rank == 0:
-                print("Measuring T1 along the polarization axis")
             self.Mz = n @ self.Mvec
         else:
             self.Mz = np.linalg.norm(self.Mvec, axis=0)
         
-        t_data = self.tlist  # Time points (ensure this is defined in your class)
-
-        tolerance = 1E-8
-
-        # Find equilibrium value when consecutive differences become smaller than threshold
-        Mz_eq = self.Mz[-1]  # Default to last value if no convergence found
+        t_data = self.tlist
         
-        for i in range(1, len(self.Mz)):
-            if abs(self.Mz[i] - self.Mz[i - 1]) < tolerance:
-                Mz_eq = self.Mz[i]
-                break
-
+        # Find where Mz drops below 10%
+        cutoff_mask = self.Mz > 0.1
+        cutoff_idx = np.sum(cutoff_mask)
+        
+        if cutoff_idx < 3:
+            if rank == 0:
+                print(f"Warning: Only {cutoff_idx} points above 1e-2. Using minimum of 3 points.")
+            cutoff_idx = min(3, len(self.Mz))
+        
+        # Cut the data
+        t_fit = t_data[:cutoff_idx]
+        Mz_fit = self.Mz[:cutoff_idx]
+        
         if rank == 0:
-            print("Equilibrium M_parallel:", Mz_eq)
-
-        # Define the T1 model function
-        def Mz_model(t, Mz_initial, Mz_eq, T1):
-            return (Mz_initial - Mz_eq) * np.exp(-t / T1) + Mz_eq
+            print(f"Using first {cutoff_idx} of {len(self.Mz)} points (Mz > 1e-2)")
+            print(f"Time range: {t_fit[0]:.4e} to {t_fit[-1]:.4e}")
+            print(f"Mz range: {Mz_fit[0]:.4f} to {Mz_fit[-1]:.4f}")
         
-        # Initial guesses for parameters [Mz_initial, Mz_eq, T1]
-        p0 = [self.Mz[0], Mz_eq, 1]  
-
-        # Define bounds (must have length 3)
-        lower = [-np.inf, -np.inf, 0.0]   # T1 must be positive
-        upper = [ np.inf,  np.inf, np.inf]
+        Mz_eq = 0.0
         
-        # Perform the fit
-        try:
-            params, covariance = curve_fit(Mz_model, t_data, self.Mz, p0=p0,bounds=(lower, upper))
-        except Exception as e:
-            print(f"Fit failed: {e}")
-            return None, None
+        # Calculate delta from equilibrium
[…]

[…]
