### 2026-04-01 cebecfa
Update math_func.py
