### 2026-04-13 #9538 @RMeli
@victorchanglee apologies for the massive delay here, I had very little time to dedicate to JOSS this month. I'll try to finalize the review ASAP.
