### 2026-04-21 #9538 @victorchanglee
@RMeli Thanks so much for your time!

- [x] Double check authors and affiliations (including ORCIDs).
- [x] Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper:
       - Updated release version with the latest changes after review.
       -  JOSS released version will be v1.0.0
       - https://doi.org/10.5281/zenodo.19682916
