+++ README.md
@@ -1,6 +1,7 @@
 [![status](https://joss.theoj.org/papers/6cd884ec29554707741cd8e700542a68/status.svg)](https://joss.theoj.org/papers/6cd884ec29554707741cd8e700542a68)
 [![GitHub release](https://img.shields.io/badge/version-1.0.0-blue)](https://github.com/victorchanglee/SpinVibe/releases)
 [![Documentation](https://img.shields.io/badge/docs-SpinVibe-blue)](https://victorchanglee.github.io/SpinVibe/)
+[![DOI](https://zenodo.org/badge/DOI/10.1103/physrevb.92.014102.svg)](https://doi.org/10.1103/physrevb.92.014102)
 
 
 # SpinVibe
