### 2026-04-21 9d8f0df
Add DOI badge to README
