### 2026-04-28 #9538 @RMeli
@victorchanglee the equations on the paper are not displayed correctly, since they are too long. Could you please reformat them, so that they are displayed appropriately? See page 3.
