### 2026-04-30 8746c76
Update paper.md
