### 2026-04-30 299670c
Improve formatting of Redfield tensor equation
Formatted the single-phonon contribution equation for clarity.
