### 2026-04-30 2cc242c
Update paper.md
