### 2026-04-30 cadecca
Update paper.md
