### 2026-04-30 d5e8952
Refactor two-phonon contribution equation formatting
