### 2026-04-30 acf3953
Fix formatting of G^{2--ph} equation in paper.md
