### 2026-05-04 0dbc2d9
Fix formatting and phrasing in paper.md
