### 2026-05-04 cc91262
Update paper.bib
