+++ docs/paper.md
@@ -192,45 +192,6 @@ grants #2138259, #2138286, #2138307, #2137603, and #2138296.
 V. C. L. thanks L. A. Mariano and A. Lunghi for the helpful discussions.
 
 # References
-## References
-
-[1] S. L. Bayliss *et al.*, *Enhancing Spin Coherence in Optically Addressable Molecular Qubits through Host-Matrix Control*,  
-**Phys. Rev. X** 12, 031028 (2022). https://doi.org/10.1103/PhysRevX.12.031028
-
-[2] M. K. Wojnar *et al.*, *Ligand field design enables quantum manipulation of spins in Ni²⁺ complexes*,  
-**Chem. Sci.** 15, 1374–1383 (2024). https://doi.org/10.1039/D3SC04919A
-
-[3] D. W. Laorenza and D. E. Freedman, *Could the Quantum Internet Be Comprised of Molecular Spins with Tunable Optical Interfaces?*,  
-**J. Am. Chem. Soc.** 144, 21810–21825 (2022). https://doi.org/10.1021/jacs.2c07775
-
-[4] M. Warner *et al.*, *Potential for spin-based information processing in a thin-film molecular semiconductor*,  
-**Nature** 503, 504–508 (2013). https://doi.org/10.1038/nature12597
-
-[5] K. R. Mullin *et al.*, *Systems-chart approach to the design of spin relaxation times in molecular qubits*,  
-**Dalton Trans.** 53, 16585–16591 (2024). https://doi.org/10.1039/D4DT02311K
-
-[6] D. W. Laorenza *et al.*, *Tunable Cr⁴⁺ Molecular Color Centers*,  
-**J. Am. Chem. Soc.** 143, 21350–21363 (2021). https://doi.org/10.1021/jacs.1c10145
-
-[7] A. Abragam and B. Bleaney,  
-*Electron Paramagnetic Resonance of Transition Ions*, Oxford University Press (2012).
-
-[8] A. Lunghi and S. Sanvito, *The Limit of Spin Lifetime in Solid-State Electronic Spins*,  
-**J. Phys. Chem. Lett.** 11, 6273–6278 (2020). https://doi.org/10.1021/acs.jpclett.0c01681
-
-[9] S. Baroni, S. de Gironcoli, A. Dal Corso, and P. Giannozzi,  
-*Phonons and related crystal properties from density-functional perturbation theory*,  
-**Rev. Mod. Phys.** 73, 515–562 (2001). https://doi.org/10.1103/RevModPhys.73.515
-
-[10] K. Janicka, A. L. Wysocki, and K. Park, *Computational Insights into Electronic Excitations, Spin–Orbit Coupling Effects, and Spin Decoherence in Cr(IV)-Based Molecular Qubits*,  
-**J. Phys. Chem. A** 126, 8007–8020 (2022). https://doi.org/10.1021/acs.jpca.2c06854
-
-[11] L. Baldinelli *et al.*, *Design Rules to Engineer the Spin Structure of Cr⁴⁺ Molecular Qubits via Matrix Modularity*,  
-**J. Am. Chem. Soc.** 147, 20693–20702 (2025). https://doi.org/10.1021/jacs.5c04004
-
-[12] Lunghi Group. (2024). MolForge (v1.0.3‑T4) [Software]. Zenodo. https://doi.org/10.5281/zenodo.14050193
-
-[13] Chilton Group. (2026). spin‑phonon‑suite (Version 1.14.1) [Computer software]. https://pypi.org/project/spin-phonon-suite/
 
 [14] A. Lunghi, *Toward exact predictions of spin-phonon relaxation times: An ab initio implementation of open quantum systems theory*,  
 **Sci. Adv.** 8, eabn7880 (2022). https://doi.org/10.1126/sciadv.abn7880
