### 2026-05-04 c6669c6
Delete docs/citations.bib
