+++ docs/citations.bib
@@ -1,214 +1,0 @@
-@Article{Lunghi2020,
-author={Lunghi, Alessandro
-and Sanvito, Stefano},
-title={The Limit of Spin Lifetime in Solid-State Electronic Spins},
-journal={The Journal of Physical Chemistry Letters},
-year={2020},
-month={Aug},
-day={06},
-publisher={American Chemical Society},
-volume={11},
-number={15},
-pages={6273-6278},
-doi={10.1021/acs.jpclett.0c01681},
-url={https://doi.org/10.1021/acs.jpclett.0c01681}
-}
-
-@article{lunghi2022,
-author = {Alessandro Lunghi },
-title = {Toward exact predictions of spin-phonon relaxation times: An ab initio implementation of open quantum systems theory},
-journal = {Science Advances},
-volume = {8},
-number = {31},
-pages = {eabn7880},
-year = {2022},
-doi = {10.1126/sciadv.abn7880},
-URL = {https://www.science.org/doi/abs/10.1126/sciadv.abn7880},
-eprint = {https://www.science.org/doi/pdf/10.1126/sciadv.abn7880},
-abstract = {Spin-phonon coupling is the main driver of spin relaxation and decoherence in solid-state semiconductors at finite temperature. Controlling this interaction is a central problem for many disciplines, ranging from magnetic resonance to quantum technologies. Spin relaxation theories have been developed for almost a century but often use a phenomenological description of phonons and their coupling to spin, resulting in a nonpredictive tool and hindering our detailed understanding of spin dynamics. Here, we combine time-local master equations up to the fourth order with advanced electronic structure methods and perform predictions of spin-phonon relaxation time for a series of solid-state coordination compounds based on both transition metals and lanthanide Kramers ions. The agreement between experiments and simulations demonstrates that an accurate, universal, and fully ab initio implementation of spin relaxation theory is possible, thus paving the way to a systematic study of spin-phonon relaxation in solid-state materials. Ab initio open quantum system theory can predict spin-phonon relaxation time in solid-state magnetic systems.}}
-
-@Inbook{Lunghi2023,
-  author    = {Lunghi, Alessandro},
-  editor    = {Rajaraman, Gopalan},
-  title     = {Spin-Phonon Relaxation in Magnetic Molecules: Theory, Predictions and Insights},
-  bookTitle = {Computational Modelling of Molecular Nanomagnets},
-  year      = {2023},
-  publisher = {Springer International Publishing},
-  address   = {Cham},
-  pages     = {219--289},
-  isbn      = {978-3-031-31038-6},
-  doi       = {10.1007/978-3-031-31038-6_6},
-  url       = {https://doi.org/10.1007/978-3-031-31038-6_6}
-}
-
-﻿@Article{Laorenza2021,
-author={Laorenza, Daniel W.
-and Kairalapova, Arailym
-and Bayliss, Sam L.
-and Goldzak, Tamar
-and Greene, Samuel M.
-and Weiss, Leah R.
-and Deb, Pratiti
-and Mintun, Peter J.
-and Collins, Kelsey A.
-and Awschalom, David D.
-and Berkelbach, Timothy C.
-and Freedman, Danna E.},
-title={{Tunable Cr$^{4+}$ Molecular Color Centers}},
-journal={Journal of the American Chemical Society},
-year={2021},
-month={Dec},
-day={22},
-publisher={American Chemical Society},
-volume={143},
-number={50},
-pages={21350-21363},
-issn={0002-7863},
[…]
