### 2026-05-04 b45b908
Update paper.bib
