+++ docs/paper.bib
@@ -11,7 +11,7 @@
   year = {2022},
   month = Aug 
 }
-@ARTICLE{Wojnar2024-nc,
+@ARTICLE{Wojnar2024,
   title    = "Ligand field design enables quantum manipulation of spins in Ni2+
               complexes",
   author   = "Wojnar, Michael K and Kundu, Krishnendu and Kairalapova, Arailym
@@ -52,7 +52,7 @@
   year     =  2024,
   language = "en"
 }
-@ARTICLE{Laorenza2022-pj,
+@ARTICLE{Laorenza2022,
   title     = "Could the quantum Internet be comprised of molecular spins with
                tunable optical interfaces?",
   author    = "Laorenza, Daniel W and Freedman, Danna E",
@@ -84,7 +84,7 @@
   language  = "en"
 }
 
-@ARTICLE{Warner2013-vf,
+@ARTICLE{Warner2013,
   title     = "Potential for spin-based information processing in a thin-film
                molecular semiconductor",
   author    = "Warner, Marc and Din, Salahud and Tupitsyn, Igor S and Morley,
@@ -126,7 +126,7 @@
   language  = "en"
 }
 
-@ARTICLE{Mullin2024-zj,
+@ARTICLE{Mullin2024,
   title     = "Systems-chart approach to the design of spin relaxation times in
                molecular qubits",
   author    = "Mullin, Kathleen R and Johnson, Dane and Freedman, Danna E and
@@ -159,7 +159,7 @@
   language  = "en"
 }
 
-@ARTICLE{Laorenza2021-tw,
+@ARTICLE{Laorenza2021,
   title     = "Tunable Cr4+ molecular color centers",
   author    = "Laorenza, Daniel W and Kairalapova, Arailym and Bayliss, Sam L
                and Goldzak, Tamar and Greene, Samuel M and Weiss, Leah R and
@@ -201,7 +201,7 @@
   copyright = "https://creativecommons.org/licenses/by-nc-nd/4.0/",
   language  = "en"
 }
-@BOOK{Abragam2012-ud,
+@BOOK{Abragam2012,
   title     = "Electron paramagnetic resonance of transition ions",
   author    = "Abragam, Anatole and Bleaney, B",
   publisher = "Oxford University Press",
@@ -210,7 +210,7 @@
   year      =  2012,
   address   = "London, England"
 }
-@ARTICLE{Lunghi2020-gq,
+@ARTICLE{Lunghi2020,
   title     = "The limit of spin lifetime in solid-state electronic spins",
   author    = "Lunghi, Alessandro and Sanvito, Stefano",
   abstract  = "The development of spin qubits for quantum technologies requires
@@ -238,7 +238,7 @@
   language  = "en"
 }
 
-@ARTICLE{Baroni2001-wx,
+@ARTICLE{Baroni2001,
   title     = "Phonons and related crystal properties from density-functional
                perturbation theory",
   author    = "Baroni, Stefano and de Gironcoli, Stefano and Dal Corso, Andrea
@@ -264,7 +264,7 @@
   language  = "en"
 }
 
-@ARTICLE{Janicka2022-wq,
+@ARTICLE{Janicka2022,
   title     = "Computational insights into electronic excitations, spin-orbit
                coupling effects, and spin decoherence in {Cr(IV)-based}
                molecular qubits",
@@ -307,7 +307,7 @@
   language  = "en"
 }
 
-@ARTICLE{Baldinelli2025-yw,
+@ARTICLE{Baldinelli2025,
   title    = "Design rules to engineer the spin structure of Cr4+ molecular
               qubits via matrix modularity",
[…]
