+++ docs/paper.md
@@ -94,7 +94,7 @@ long-lived spin states.
 To this end, we developed `SpinVibe`, a Python package that is designed
 to be user-friendly, providing streamlined workflows that are readily
 adapted to any *ab initio* code. Although the software packages MolForge [@Lunghiale2024] 
-and Spin-phonon suite [@PyPI] have been developed to address this problem, 
+and Spin-phonon suite [@spin-phonon-suite] have been developed to address this problem, 
 SpinVibe is intentionally designed for user flexibility and increased interoperability. 
 Our code uses a Python based framework, allowing the storage of the required data from 
 first-principles calculations as `NumPy` arrays in HDF5 files, and is not restricted to 
