### 2026-05-04 e6e93b8
Add spin-phonon-suite reference to bibliography
