+++ docs/paper.bib
@@ -347,7 +347,13 @@
   publisher = "Zenodo",
   year      =  2024
 }
-@misc{PyPI, url={https://pypi.org/project/spin-phonon-suite/}, journal={PyPI}} 
+@misc{spin-phonon-suite,
+	title = {spin-phonon-suite: {A} package for performing spin-phonon coupling calculations with {openMOLCAS}, {VASP}, and {Gaussian}},
+	copyright = {OSI Approved :: GNU General Public License v3},
+	shorttitle = {spin-phonon-suite},
+	url = {https://gitlab.com/chilton-group/spin_phonon_suite},
+	urldate = {2026-05-04},
+}
 @INCOLLECTION{Lunghi2023-em,
   title     = "Spin-phonon relaxation in magnetic molecules: Theory,
                predictions and insights",
