### 2026-05-04 d02c873
Update paper.bib
