### 2026-05-04 b6b5900
Update paper.md
