### 2026-05-04 0589448
Fix citation year in Lunghi reference
