+++ docs/paper.bib
@@ -354,7 +354,7 @@
 	url = {https://gitlab.com/chilton-group/spin_phonon_suite},
 	urldate = {2026-05-04},
 }
-@INCOLLECTION{Lunghi:20230,
+@INCOLLECTION{Lunghi:2023,
   title     = "Spin-phonon relaxation in magnetic molecules: Theory,
                predictions and insights",
   booktitle = "Challenges and Advances in Computational Chemistry and Physics",
