### 2026-05-04 af5653f
Update Spin-phonon-suite citation details
