+++ docs/paper.bib
@@ -347,10 +347,10 @@
   publisher = "Zenodo",
   year      =  2024
 }
-@misc{Spin-phonon-suite,
+@misc{Spin-phonon-suite:2026,
 	title = {spin-phonon-suite: {A} package for performing spin-phonon coupling calculations with {openMOLCAS}, {VASP}, and {Gaussian}},
 	copyright = {OSI Approved :: GNU General Public License v3},
-	shorttitle = {spin-phonon-suite},
+	shorttitle = {Spin-phonon-suite},
 	url = {https://gitlab.com/chilton-group/spin_phonon_suite},
 	urldate = {2026-05-04},
 }
