### 2026-05-04 982ad67
Fix citation year for Spin-phonon suite
Updated references for Spin-phonon suite citation.
