### 2026-05-04 d64f2d2
Update paper.md
