+++ docs/paper.bib
@@ -352,6 +352,7 @@
 	copyright = {OSI Approved :: GNU General Public License v3},
 	shorttitle = {Spin-phonon-suite},
 	url = {https://gitlab.com/chilton-group/spin_phonon_suite},
+ 	year      =  2026,
 	urldate = {2026-05-04},
 }
 @INCOLLECTION{Lunghi:2023,
