+++ docs/paper.bib
@@ -11,47 +11,19 @@
   year = {2022},
   month = Aug 
 }
-@ARTICLE{Wojnar:2024,
-  title    = "Ligand field design enables quantum manipulation of spins in Ni2+
-              complexes",
-  author   = "Wojnar, Michael K and Kundu, Krishnendu and Kairalapova, Arailym
-              and Wang, Xiaoling and Ozarowski, Andrew and Berkelbach, Timothy
-              C and Hill, Stephen and Freedman, Danna E",
-  abstract = "Creating the next generation of quantum systems requires control
-              and tunability, which are key features of molecules. To design
-              these systems, one must consider the ground-state and
-              excited-state manifolds. One class of systems with promise for
-              quantum sensing applications, which require water solubility, are
-              d8 Ni2+ ions in octahedral symmetry. Yet, most Ni2+ complexes
-              feature large zero-field splitting, precluding manipulation by
-              commercial microwave sources due to the relatively large
-              spin-orbit coupling constant of Ni2+ (630 cm-1). Since low lying
-              excited states also influence axial zero-field splitting, D, a
-              combination of strong field ligands and rigidly held octahedral
-              symmetry can ameliorate these challenges. Towards these ends, we
-              performed a theoretical and computational analysis of the
-              electronic and magnetic structure of a molecular qubit, focusing
-              on the impact of ligand field strength on D. Based on those
-              results, we synthesized 1, [Ni(ttcn)2](BF4)2 (ttcn =
-              1,4,7-trithiacyclononane), which we computationally predict will
-              have a small D (Dcalc = +1.15 cm-1). High-field high-frequency
-              electron paramagnetic resonance (EPR) data yield spin Hamiltonian
-              parameters: gx = 2.1018(15), gx = 2.1079(15), gx = 2.0964(14), D
-              = +0.555(8) cm-1 and E = +0.072(5) cm-1, which confirm the
-              expected weak zero-field splitting. Dilution of 1 in the
-              diamagnetic Zn analogue, [Ni0.01Zn0.99(ttcn)2](BF4)2 (1') led to
-              a slight increase in D to ∼0.9 cm-1. The design criteria in
-              minimizing D in 1via combined computational and experimental
-              methods demonstrates a path forward for EPR and optical
-              addressability of a general class of S = 1 spins.",
-  journal  = "Chem. Sci.",
-  volume   =  15,
-  number   =  4,
-  pages    = "1374--1383",
-  month    =  jan,
-  year     =  2024,
-  language = "en"
-}
+@Article{Wojnar:2024,
+author ="Wojnar, Michael K. and Kundu, Krishnendu and Kairalapova, Arailym and Wang, Xiaoling and Ozarowski, Andrew and Berkelbach, Timothy C. and Hill, Stephen and Freedman, Danna E.",
+title  ="Ligand field design enables quantum manipulation of spins in Ni2+ complexes",
+journal  ="Chem. Sci.",
+year  ="2024",
+volume  ="15",
+issue  ="4",
+pages  ="1374-1383",
+publisher  ="The Royal Society of Chemistry",
+doi  ="10.1039/D3SC04919A",
+url  ="http://dx.doi.org/10.1039/D3SC04919A",
[…]
