### 2026-05-04 0400b8d
Update paper.bib
