+++ docs/paper.bib
@@ -24,78 +24,68 @@ doi  ="10.1039/D3SC04919A",
 url  ="http://dx.doi.org/10.1039/D3SC04919A",
 abstract  ="Creating the next generation of quantum systems requires control and tunability{,} which are key features of molecules. To design these systems{,} one must consider the ground-state and excited-state manifolds. One class of systems with promise for quantum sensing applications{,} which require water solubility{,} are d8 Ni2+ ions in octahedral symmetry. Yet{,} most Ni2+ complexes feature large zero-field splitting{,} precluding manipulation by commercial microwave sources due to the relatively large spin–orbit coupling constant of Ni2+ (630 cm−1). Since low lying excited states also influence axial zero-field splitting{,} D{,} a combination of strong field ligands and rigidly held octahedral symmetry can ameliorate these challenges. Towards these ends{,} we performed a theoretical and computational analysis of the electronic and magnetic structure of a molecular qubit{,} focusing on the impact of ligand field strength on D. Based on those results{,} we synthesized 1{,} [Ni(ttcn)2](BF4)2 (ttcn = 1{,}4{,}7-trithiacyclononane){,} which we computationally predict will have a small D (Dcalc = +1.15 cm−1). High-field high-frequency electron paramagnetic resonance (EPR) data yield spin Hamiltonian parameters: gx = 2.1018(15){,} gx = 2.1079(15){,} gx = 2.0964(14){,} D = +0.555(8) cm−1 and E = +0.072(5) cm−1{,} which confirm the expected weak zero-field splitting. Dilution of 1 in the diamagnetic Zn analogue{,} [Ni0.01Zn0.99(ttcn)2](BF4)2 (1′) led to a slight increase in D to ∼0.9 cm−1. The design criteria in minimizing D in 1via combined computational and experimental methods demonstrates a path forward for EPR and optical addressability of a general class of S = 1 spins."}
 
-@ARTICLE{Laorenza:2022,
-  title     = "Could the quantum Internet be comprised of molecular spins with
-               tunable optical interfaces?",
-  author    = "Laorenza, Daniel W and Freedman, Danna E",
-  abstract  = "One goal of the second quantum revolution is developing an
-               approach to transmit coherent quantum information across the
-               world, forming the basis of the quantum Internet. Achieving this
-               lofty goal will require scalable and tunable materials that
-               allow for cross-platform interconnections. Solution-processable
-               molecular systems designed from the ground up are poised to
-               introduce the necessary tunability to meet these needs. In this
-               Perspective, we describe the components of a quantum network,
-               outline the criteria to create molecular systems suitable for
-               networking applications, and discuss the pathway to generate
-               spin-photon entanglement, a first step toward remote
-               entanglement between molecular spins in a network. We also
-               describe some initial steps to optimize molecular spins for
-               potential integration into quantum networks as readily
-               deployable qubits in network nodes. Throughout this Perspective,
-               we highlight the tremendous potential of molecular systems for
-               the quantum Internet and illustrate openings for chemists in
-               this nascent field.",
-  journal   = "J. Am. Chem. Soc.",
-  publisher = "American Chemical Society (ACS)",
-  volume    =  144,
-  number    =  48,
-  pages     = "21810--21825",
-  month     =  dec,
-  year      =  2022,
-  language  = "en"
+@Inbook{Zhang:2002,
+author="Zhang, Guoping
+and H{\"u}bner, Wolfgang
+and Beaurepaire, Eric
+and Bigot, Jean-Yves",
+editor="Hillebrands, Burkard
+and Ounadjela, Kamel",
+title="Laser-Induced Ultrafast Demagnetization: Femtomagnetism, a New Frontier?",
[…]
