+++ docs/paper.bib
@@ -212,7 +212,7 @@ url={https://doi.org/10.1021/jacs.5c04004}
                of spin dynamics at the atomistic level. The code is developed
                to readily interface with first-principles codes.",
   publisher = "Zenodo",
-  doi={https://doi.org/10.5281/zenodo.6812141},
+  doi={10.5281/zenodo.6812141},
   year      =  2024
 }
 @misc{Spin-phonon-suite:2026,
