### 2026-05-04 cb30c71
Update paper.bib
