+++ docs/paper.bib
@@ -60,9 +60,7 @@ doi={10.1021/jacs.2c07775},
 url={https://doi.org/10.1021/jacs.2c07775}
 }
 
-
-
-@ARTICLE{Warner:2013,
+@Article{Warner:2013,
 author={Warner, Marc
 and Din, Salahud
 and Tupitsyn, Igor S.
@@ -87,6 +85,8 @@ issn={1476-4687},
 doi={10.1038/nature12597},
 url={https://doi.org/10.1038/nature12597}
 }
+
+
 
 @ARTICLE{Mullin:2024,
 author ="Mullin, Kathleen R. and Johnson, Dane and Freedman, Danna E. and Rondinelli, James M.",
