### 2026-05-04 272f899
Fix citation formatting in paper.md
