### 2026-05-05 477c16c
Update paper.md
